@@ -9,84 +9,67 @@ exclude = ["local_ignore"]
 module = "cloudpickle.*,fiona.*,geopandas.*,matplotlib.*,osgeo.*,osgeo_utils.*,pygeos.*,pyogrio.*,psutil.*,setuptools.*,shapely.*,topojson.*"
 ignore_missing_imports = true
 
+[tool.pytest.ini_options]
+log_level = "DEBUG"
+# Write all logging and warnings immediately. Useful for debugging.
+# log_cli = true
+# addopts = "-p no:warnings"
+
+# Some filters that apply to warnings that can be ignored for any test.
+filterwarnings = [
+    "ignore: The output driver does not natively support DateTime type",
+    "ignore: Field .* create as date field, though DateTime requested",
+    "ignore: Geometry is in a geographic CRS",
+    "ignore: Normalized/laundered field name",
+    "ignore: Non-conformant content for record 0 in column",
+    "ignore: Non-conformant content for record 1 in column",
+    "ignore: Using GFO_IO_ENGINE=fiona is deprecated",
+]
+
 [tool.ruff]
 line-length = 88
+target-version = "py310"
+extend-exclude = ["docs/*", "local_ignore/*"]
+
+[tool.ruff.lint]
 select = [
-    # pyflakes
-    "F",
-    # pycodestyle
-    "E",
+    "F",  # pyflakes
+    "E",  # pycodestyle
     "W",
-    # flake8-2020
-    "YTT",
-    # flake8-bugbear
-    "B",
-    # flake8-quotes
-    "Q",
-    # flake8-debugger
-    "T10",
-    # flake8-gettext
-    "INT",
-    # pylint
-    "PLC",
+    "UP",  # pyupgrade
+    "B",  # flake8-bugbear
+    "PTH",  # flake8-pathlib
+    "T10",  # flake8-debugger
+    # "SIM",  # flake8-simplify
+    "PLC",  # pylint
     "PLE",
     "PLR",
     "PLW",
-    # misc lints
-    "PIE",
-    # flake8-pyi
-    "PYI",
-    # tidy imports
-    "TID",
-    # implicit string concatenation
-    "ISC",
-    # type-checking imports
-    "TCH",
-    # comprehensions
-    "C4",
-    # pygrep-hooks
-    "PGH",
-    # Ruff-specific rules
-    "RUF",
-    # pyupgrade
-    "UP",
-    # pydocstyle
-    "D",
+    "PIE",  # misc lints
+    "ISC",  # implicit string concatenation
+    "TCH",  # type-checking imports
+    "C4",  # comprehensions
+    "RUF",  # Ruff-specific rules
+    "I",  # isort
+    "D",  # pydocstyle
 ]
 
-target-version = "py38"
-ignore = [ # space before : (needed for how black formats slicing)
-    # "E203",  # not yet implemented
+ignore = [
+    ### Intentionally disabled
+    # module level import not at top of file
+    "E402",
     # do not assign a lambda expression, use a def
     "E731",
-    # line break before binary operator
-    # "W503",  # not yet implemented
-    # line break after binary operator
-    # "W504",  # not yet implemented
-    # controversial
+    # mutable-argument-default
     "B006",
-    # controversial
+    # unused-loop-control-variable
     "B007",
-    # controversial
-    "B008",
-    # setattr is used to side-step mypy
+    # get-attr-with-constant
     "B009",
-    # getattr is used to side-step mypy
-    "B010",
-    # tests use assert False
-    "B011",
-    # tests use comparisons but not their returned value
-    "B015",
-    # false positives
-    "B019",
-    # Loop control variable overrides iterable it iterates
-    "B020",
-    # Function definition does not bind loop variable
-    "B023",
-    # Functions defined inside a loop must not use variables redefined in the loop
-    # "B301",  # not yet implemented
     # Only works with python >=3.10
     "B905",
+    # dict literals
+    "C408",
     # Too many arguments to function call
     "PLR0913",
     # Too many returns
