@@ -1,13 +1,13 @@
-"""
-Setup file to package geofileops.
-"""
+"""Setup file to package geofileops."""
+
+from pathlib import Path
 
 import setuptools
 
-with open("README.md") as fh:
+with Path("README.md").open() as fh:
     long_description = fh.read()
 
-with open("geofileops/version.txt") as file:
+with Path("geofileops/version.txt").open() as file:
     version = file.readline()
 
 setuptools.setup(
@@ -20,25 +20,31 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     url="https://github.com/geofileops/geofileops",
     include_package_data=True,
-    packages=setuptools.find_packages(),
+    packages=setuptools.find_packages(
+        exclude=("tests", "benchmark", "benchmark.*"),
+        include=("geofileops", "geofileops.*"),
+    ),
     install_requires=[
         "cloudpickle",
-        "fiona",
         "gdal>=3.6",
-        "geopandas>=0.12,<1",
+        "geopandas>=0.12",
         "numpy",
         "packaging",
-        "pandas",
+        "pandas>=1.5",
         "psutil",
-        "pygeoops>=0.4,<0.5",
-        "pyogrio",
+        "pygeoops>=0.4",
+        "pyogrio>=0.7",
         "pyproj",
-        "shapely>=2,<2.1",
+        "shapely>=2",
     ],
     extras_require={"full": ["simplification"]},
     classifiers=[
-        "Programming Language :: Python :: 3",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: BSD License",
         "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3 :: Only",
+        "Topic :: Scientific/Engineering :: GIS",
     ],
-    python_requires=">=3.8",
+    python_requires=">=3.10",
 )
