@@ -1,14 +1,11 @@
-"""
-Module for benchmarking.
-"""
+"""Module for benchmarking."""
 
 import datetime
 import importlib
 import inspect
 import logging
-from pathlib import Path
 import tempfile
-from typing import List, Optional
+from pathlib import Path
 
 import pandas as pd
 
@@ -35,10 +32,9 @@ class RunResult:
         operation: str,
         operation_descr: str,
         secs_taken: float,
-        run_details: Optional[dict] = None,
+        run_details: dict | None = None,
     ):
-        """
-        Constructor for a RunResult.
+        """Constructor for a RunResult.
 
         Args:
             package (str): Package being benchmarked.
@@ -58,18 +54,15 @@ class RunResult:
         self.run_details = run_details
 
     def __repr__(self):
-        """
-        Format the result.
-        """
+        """Format the result."""
         return f"{self.__class__}({self.__dict__})"
 
 
 def run_benchmarks(
-    modules_to_run: Optional[List[str]] = None,
-    functions_to_run: Optional[List[str]] = None,
+    modules_to_run: list[str] | None = None,
+    functions_to_run: list[str] | None = None,
 ):
-    """
-    Run all benchmarks specified.
+    """Run all benchmarks specified.
 
     Args:
         modules_to_run (Optional[List[str]], optional): List of modules to run
