@@ -1,13 +1,11 @@
-"""
-Module to generate reports for benchmarks.
-"""
+"""Module to generate reports for benchmarks."""
 
 import ast
 import math
-from pathlib import Path
 import shutil
 import tempfile
-from typing import Literal, Optional, Tuple
+from pathlib import Path
+from typing import Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -20,8 +18,7 @@ A4_SHORT_SIDE = 8.27
 
 
 def generate_reports(results_path: Path, output_dir: Path):
-    """
-    Generate the benchmarking reports.
+    """Generate the benchmarking reports.
 
     Args:
         results_path (Path): the result path to find the results to report on.
@@ -30,7 +27,7 @@ def generate_reports(results_path: Path, output_dir: Path):
     benchmark_df = pd.read_csv(results_path)
 
     def format_run_details(input: dict) -> str:
-        if input is None or input == np.nan:
+        if input is None or input == np.nan:  # noqa: PLW0177
             return ""
         if isinstance(input, str):
             input = ast.literal_eval(input)
@@ -112,10 +109,10 @@ def save_chart(
     df: pd.DataFrame,
     title: str,
     output_path: Path,
-    yscale: Optional[Literal["linear", "log", "symlog", "logit"]] = None,
-    y_value_formatter: Optional[str] = None,
+    yscale: Literal["linear", "log", "symlog", "logit"] | None = None,
+    y_value_formatter: str | None = None,
     print_labels_on_points: bool = False,
-    size: Tuple[float, float] = (8, 4),
+    size: tuple[float, float] = (8, 4),
     plot_kind: Literal[
         "line",
         "bar",
@@ -129,11 +126,10 @@ def save_chart(
         "scatter",
         "hexbin",
     ] = "line",
-    gridlines: Optional[Literal["both", "x", "y"]] = None,
-    linestyle: Optional[str] = None,
+    gridlines: Literal["both", "x", "y"] | None = None,
+    linestyle: str | None = None,
 ):
-    """
-    Render and save a chart.
+    """Render and save a chart.
 
     Args:
         df (pd.DataFrame): _description_
@@ -203,7 +199,7 @@ def save_chart(
 
         label_above_line = True
         for index, row in enumerate(df.itertuples()):
-            for row_fieldname, row_fieldvalue in row._asdict().items():
+            for row_fieldname, row_fieldvalue in row._asdict().items():  # type: ignore
                 if row_fieldname != "Index":
                     if max_y_value is None or row_fieldvalue > max_y_value:
                         max_y_value = row_fieldvalue
@@ -250,7 +246,7 @@ def save_chart(
     if not output_path.exists():
         fig.savefig(str(output_path))
     else:
-        # If it exists already, only save it if it has changed
+        # If it already exists, only save it if it has changed
         with tempfile.TemporaryDirectory() as tmp_dir:
             tmp_output_path = Path(tmp_dir) / output_path.name
             fig.savefig(tmp_output_path)
@@ -258,6 +254,8 @@ def save_chart(
             img_old = np.asarray(Image.open(output_path))
             if not np.array_equal(img_new, img_old):
                 shutil.move(tmp_output_path, output_path)
+
+    plt.close(fig=fig)
 
 
 if __name__ == "__main__":
