@@ -1,33 +1,34 @@
-"""
-Module exposing all supported operations on geometries in geofiles.
-"""
+"""Module exposing all supported operations on geometries in geofiles."""
 
-from datetime import datetime
 import logging
 import logging.config
-from pathlib import Path
 import shutil
-from typing import Any, Callable, List, Literal, Optional, Tuple, Union, TYPE_CHECKING
 import warnings
+from collections.abc import Callable
+from datetime import datetime
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Literal, Union
 
 from pygeoops import GeometryType
 
-from geofileops._compat import SPATIALITE_GTE_51
 from geofileops import fileops
+from geofileops._compat import SPATIALITE_GTE_51
 from geofileops.helpers._configoptions_helper import ConfigOptions
-from geofileops.util import _geofileinfo
-from geofileops.util import _geoops_gpd
-from geofileops.util import _geoops_sql
-from geofileops.util import _geoops_ogr
-from geofileops.util import _io_util
-from geofileops.util import _sqlite_util
+from geofileops.util import (
+    _geofileinfo,
+    _geoops_gpd,
+    _geoops_ogr,
+    _geoops_sql,
+    _io_util,
+    _sqlite_util,
+)
 from geofileops.util._geometry_util import (
     BufferEndCapStyle,
     BufferJoinStyle,
     SimplifyAlgorithm,
 )
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     import os
 
 logger = logging.getLogger(__name__)
@@ -39,14 +40,13 @@ def dissolve_within_distance(
     distance: float,
     gridsize: float,
     close_internal_gaps: bool = False,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Dissolve geometries that are within the distance specified.
+    """Dissolve geometries that are within the distance specified.
 
     The output layer will contain the dissolved geometries where all gaps between the
     input geometries up to ``distance`` are closed.
@@ -78,10 +78,10 @@ def dissolve_within_distance(
         close_internal_gaps (bool, optional): also close gaps, strips or holes in the
             input geometries that are narrower than the ``distance`` specified. E.g.
             small holes, narrow strips starting at the boundary,... Defaults to False.
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         nb_parallel (int, optional): the number of parallel processes to use.
             Defaults to -1: use all available CPUs.
         batchsize (int, optional): indicative number of rows to process per
@@ -90,27 +90,27 @@ def dissolve_within_distance(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`dissolve`: dissolve the input layer
+
     """
     input_path = Path(input_path)
     output_path = Path(output_path)
+    if _io_util.output_exists(path=output_path, remove_if_exists=force):
+        return
 
     start_time = datetime.now()
     operation_name = "dissolve_within_distance"
     logger = logging.getLogger(f"geofileops.{operation_name}")
     nb_steps = 9
 
-    # Already check here if it is useful to continue
-    if _io_util.output_exists(path=output_path, remove_if_exists=force):
-        return
-
     tempdir = _io_util.create_tempdir(f"geofileops/{operation_name}")
     try:
         # First dissolve the input.
         #
         # Note: this reduces the complexity of operations to be executed later on.
-        # Note2: this already applies the gridsize, which needs to be applied anyway to
-        # avoid issues when determining the addedpieces_1neighbour later on.
-        # Note2: don' apply gridsize yet
+        # Note2: don't apply gridsize yet
         logger.info(f"Start, with input file {input_path}")
         step = 1
         logger.info(f"Step {step} of {nb_steps}")
@@ -203,9 +203,9 @@ def dissolve_within_distance(
         # temporary boundariesstep += 1
         logger.info(f"Step {step} of {nb_steps}")
         parts_to_add_path = tempdir / "200_parts_to_add.gpkg"
-        _geoops_sql.erase(
-            input_path=bufp_diss_bufm_path,
-            erase_path=diss_path,
+        _geoops_sql.difference(
+            input1_path=bufp_diss_bufm_path,
+            input2_path=diss_path,
             output_path=parts_to_add_path,
             overlay_self=False,
             explodecollections=True,
@@ -252,7 +252,7 @@ def dissolve_within_distance(
             #   - if > 1 neighbour, seems OK.
             #
             # For all pieces that don't comply to the above, the following parameters
-            # indicate that they need to be selected to erase them:
+            # indicate that they need to be selected to difference them:
             #   - pieces can be very narrow slivers. E.g. alongside a long boundary with
             #     a small bend, probably due to rounding side effects in the +/- buffer.
             #   - pieces can be spikes. E.g. when a "road" of ~ 'distance' width is not
@@ -343,6 +343,7 @@ def dissolve_within_distance(
             output_path=parts_to_add_filtered_path,
             sql_stmt=sql_stmt,
             input2_layer=input_layer,
+            explodecollections=True,
             gridsize=0.0,
             nb_parallel=nb_parallel,
             batchsize=batchsize,
@@ -355,8 +356,11 @@ def dissolve_within_distance(
         step += 1
         logger.info(f"Step {step} of {nb_steps}")
         dst_layer = fileops.get_only_layer(diss_path)
-        fileops.append_to(
-            src=parts_to_add_filtered_path, dst=diss_path, dst_layer=dst_layer
+        fileops.copy_layer(
+            src=parts_to_add_filtered_path,
+            dst=diss_path,
+            dst_layer=dst_layer,
+            write_mode="append",
         )
 
         step += 1
@@ -377,7 +381,7 @@ def dissolve_within_distance(
         if ConfigOptions.remove_temp_files:
             shutil.rmtree(tempdir, ignore_errors=True)
 
-    logger.info(f"Ready, took {datetime.now()-start_time}")
+    logger.info(f"Ready, took {datetime.now() - start_time}")
 
 
 def apply(
@@ -385,22 +389,24 @@ def apply(
     output_path: Union[str, "os.PathLike[Any]"],
     func: Callable[[Any], Any],
     only_geom_input: bool = True,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Union[GeometryType, str, None] = None,
+    force_output_geometrytype: GeometryType | str | None = None,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Apply a python lambda function on the geometry column of the input file.
+    """Apply a python function on the geometry column of the input file.
 
     The result is written to the output file specified.
+
+    If the function you want to apply accepts an array of geometries as input, you can
+    typically use :func:`apply_vectorized` instead, which is faster.
 
     If ``explodecollections`` is False and the input and output file type is GeoPackage,
     the fid will be preserved. In other cases this will typically not be the case.
@@ -412,10 +418,10 @@ def apply(
         only_geom_input (bool, optional): If True, only the geometry
             column is available. If False, the entire row is input.
             Remark: when False, the operation is 50% slower. Defaults to True.
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -442,16 +448,18 @@ def apply(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
 
+    See Also:
+        * :func:`apply_vectorized`: apply a vectorized python function on the geometry
+          column
+
     Examples:
         This example shows the basic usage of ``gfo.apply``:
 
         .. code-block:: python
 
-            import geofileops as gfo
-
             gfo.apply(
-                input_path=...,
-                output_path=...,
+                input_path="input.gpkg",
+                output_path="output.gpkg",
                 func=lambda geom: pygeoops.remove_inner_rings(geom, min_area_to_keep=1),
             )
 
@@ -460,11 +468,9 @@ def apply(
 
         .. code-block:: python
 
-            import geofileops as gfo
-
             gfo.apply(
-                input_path=...,
-                output_path=...,
+                input_path="input.gpkg",
+                output_path="output.gpkg",
                 func=lambda row: pygeoops.remove_inner_rings(
                     row.geometry, min_area_to_keep=row.min_area_to_keep
                 ),
@@ -499,6 +505,110 @@ def apply(
     )
 
 
+def apply_vectorized(
+    input_path: Union[str, "os.PathLike[Any]"],
+    output_path: Union[str, "os.PathLike[Any]"],
+    func: Callable[[Any], Any],
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
+    explodecollections: bool = False,
+    force_output_geometrytype: GeometryType | str | None = None,
+    gridsize: float = 0.0,
+    keep_empty_geoms: bool = False,
+    where_post: str | None = None,
+    nb_parallel: int = -1,
+    batchsize: int = -1,
+    force: bool = False,
+):
+    """Apply a vectorized python function on the geometry column of the input file.
+
+    The result is written to the output file specified.
+
+    It is not possible to use the contents of other columns in the input file in the
+    python function. If you need this, use :func:`apply` instead.
+
+    If ``explodecollections`` is False and the input and output file type is GeoPackage,
+    the fid will be preserved. In other cases this will typically not be the case.
+
+    Args:
+        input_path (PathLike): the input file
+        output_path (PathLike): the file to write the result to
+        func (Callable): vectorized lambda function to apply to the geometry column.
+            Vectorized means here that the function should accept a shapely geometry
+            array as input and will return a shapely geometry for each item in the input
+            array.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
+        columns (List[str], optional): list of columns to retain. If None, all standard
+            columns are retained. In addition to standard columns, it is also possible
+            to specify "fid", a unique index available in all input files. Note that the
+            "fid" will be aliased eg. to "fid_1". Defaults to None.
+        explodecollections (bool, optional): True to output only simple geometries.
+            Defaults to False.
+        force_output_geometrytype (GeometryType, optional): The output geometry type to
+            force. If None, a best-effort guess is made and will always result in a
+            multi-type. Defaults to None.
+        gridsize (float, optional): the size of the grid the coordinates of the ouput
+            will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
+            the precision. Defaults to 0.0.
+        keep_empty_geoms (bool, optional): True to keep rows with empty/null geometries
+            in the output. Defaults to False.
+        where_post (str, optional): SQL filter to apply after all other processing,
+            including e.g. ``explodecollections``. It should be in sqlite syntax and
+            |spatialite_reference_link| functions can be used. Defaults to None.
+        nb_parallel (int, optional): the number of parallel processes to use.
+            Defaults to -1: use all available CPUs.
+        batchsize (int, optional): indicative number of rows to process per
+            batch. A smaller batch size, possibly in combination with a
+            smaller ``nb_parallel``, will reduce the memory usage.
+            Defaults to -1: (try to) determine optimal size automatically.
+        force (bool, optional): overwrite existing output file(s).
+            Defaults to False.
+
+    See Also:
+        * :func:`apply`: apply a python function on the geometry column
+
+    Examples:
+        This example shows the usage of ``gfo.apply_vectorized``:
+
+        .. code-block:: python
+
+            gfo.apply_vectorized(
+                input_path="input.gpkg",
+                output_path="output.gpkg",
+                func=lambda geom: pygeoops.centerline(geom, densify_distance=0),
+            )
+
+
+    .. |spatialite_reference_link| raw:: html
+
+        <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
+
+    """  # noqa: E501
+    logger = logging.getLogger("geofileops.apply_vectorized")
+    logger.info(f"Start on {input_path}")
+
+    return _geoops_gpd.apply_vectorized(
+        input_path=Path(input_path),
+        output_path=Path(output_path),
+        func=func,
+        input_layer=input_layer,
+        output_layer=output_layer,
+        columns=columns,
+        explodecollections=explodecollections,
+        force_output_geometrytype=force_output_geometrytype,
+        gridsize=gridsize,
+        keep_empty_geoms=keep_empty_geoms,
+        where_post=where_post,
+        nb_parallel=nb_parallel,
+        batchsize=batchsize,
+        force=force,
+    )
+
+
 def buffer(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
@@ -508,19 +618,18 @@ def buffer(
     join_style: BufferJoinStyle = BufferJoinStyle.ROUND,
     mitre_limit: float = 5.0,
     single_sided: bool = False,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Applies a buffer operation on geometry column of the input file.
+    """Applies a buffer operation on geometry column of the input file.
 
     The result is written to the output file specified.
 
@@ -556,10 +665,10 @@ def buffer(
             if distance is negative, the left side, if distance is positive,
             the right hand side. Only relevant for line geometries.
             Defaults to False.
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -721,15 +830,14 @@ def buffer(
 def clip_by_geometry(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    clip_geometry: Union[Tuple[float, float, float, float], str],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    clip_geometry: tuple[float, float, float, float] | str,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
-    """
-    Clip all geometries in the imput file by the geometry provided.
+    """Clip all geometries in the input file by the geometry provided.
 
     If ``explodecollections`` is False and the input and output file type is GeoPackage,
     the fid will be preserved. In other cases this will typically not be the case.
@@ -739,10 +847,10 @@ def clip_by_geometry(
         output_path (PathLike): the file to write the result to
         clip_geometry (Union[Tuple[float, float, float, float], str]): the bounds
             or WKT geometry to clip with.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -751,6 +859,10 @@ def clip_by_geometry(
             Defaults to False.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`clip`: clip geometries by the features in another layer
+
     """
     logger = logging.getLogger("geofileops.clip_by_geometry")
     logger.info(f"Start, on {input_path}")
@@ -769,19 +881,18 @@ def clip_by_geometry(
 def convexhull(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Applies a convexhull operation on the input file.
+    """Applies a convexhull operation on the input file.
 
     The result is written to the output file specified.
 
@@ -791,10 +902,10 @@ def convexhull(
     Args:
         input_path (PathLike): the input file
         output_path (PathLike): the file to write the result to
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -845,16 +956,29 @@ def convexhull(
 def delete_duplicate_geometries(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
+    priority_column: str | None = None,
+    priority_ascending: bool = True,
     explodecollections: bool = False,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
+    nb_parallel: int = -1,
+    batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Copy all rows to the output file, except for duplicate geometries.
+    """Copy all rows to the output file, except for duplicate geometries.
+
+    The check for duplicates is done using ``ST_Equals``. ``ST_Equals`` is ``True`` if`
+    the given geometries are "topologically equal". This means that the geometries have
+    the same dimension and their point-sets occupy the same space. This means e.g. that
+    the order of vertices may be different, starting points of rings can be different
+    and polygons can contain extra points if they don't change the surface occupied.
+
+    If a ``priority_column`` is specified, the row with the lowest value in this column
+    is retained. If ``priority_ascending`` is False, the row with the highest value is
+    retained.
 
     If ``explodecollections`` is False and the input and output file type is GeoPackage,
     the fid will be preserved. In other cases this will typically not be the case.
@@ -862,14 +986,18 @@ def delete_duplicate_geometries(
     Args:
         input_path (PathLike): the input file
         output_path (PathLike): the file to write the result to
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
             "fid" will be aliased eg. to "fid_1". Defaults to None.
+        priority_column (str, optional): column to use as priority for keeping rows.
+            Defaults to None.
+        priority_ascending (bool, optional): True to keep the row with the lowest
+            priority value. Defaults to True.
         explodecollections (bool, optional): True to output only simple geometries.
             Defaults to False.
         keep_empty_geoms (bool, optional): True to keep rows with empty/null geometries
@@ -877,6 +1005,12 @@ def delete_duplicate_geometries(
         where_post (str, optional): SQL filter to apply after all other processing,
             including e.g. ``explodecollections``. It should be in sqlite syntax and
             |spatialite_reference_link| functions can be used. Defaults to None.
+        nb_parallel (int, optional): the number of parallel processes to use.
+            Defaults to -1: use all available CPUs.
+        batchsize (int, optional): indicative number of rows to process per
+            batch. A smaller batch size, possibly in combination with a
+            smaller ``nb_parallel``, will reduce the memory usage.
+            Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
 
@@ -894,9 +1028,13 @@ def delete_duplicate_geometries(
         input_layer=input_layer,
         output_layer=output_layer,
         columns=columns,
+        priority_column=priority_column,
+        priority_ascending=priority_ascending,
         explodecollections=explodecollections,
         keep_empty_geoms=keep_empty_geoms,
         where_post=where_post,
+        nb_parallel=nb_parallel,
+        batchsize=batchsize,
         force=force,
     )
 
@@ -905,20 +1043,19 @@ def dissolve(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     explodecollections: bool,
-    groupby_columns: Union[List[str], str, None] = None,
-    agg_columns: Optional[dict] = None,
+    groupby_columns: list[str] | str | None = None,
+    agg_columns: dict | None = None,
     tiles_path: Union[str, "os.PathLike[Any]", None] = None,
     nb_squarish_tiles: int = 1,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Applies a dissolve operation on the input file.
+    """Applies a dissolve operation on the input file.
 
     If columns are specified with ``groupby_columns``, the data is first grouped
     on those columns before the geometries are merged.
@@ -935,11 +1072,9 @@ def dissolve(
 
     .. code-block:: python
 
-        import geofileops as gfo
-
         gfo.dissolve(
-            input_path=...,
-            output_path=...,
+            input_path="input.gpkg",
+            output_path="output.gpkg",
             groupby_columns=["cropgroup"],
             agg_columns={
                 "columns": [
@@ -966,11 +1101,9 @@ def dissolve(
 
     .. code-block:: python
 
-        import geofileops as gfo
-
         gfo.dissolve(
-            input_path=...,
-            output_path=...,
+            input_path="input.gpkg",
+            output_path="output.gpkg",
             groupby_columns=["cropgroup"],
             agg_columns={"json": ["crop", "area"]},
             explodecollections=False,
@@ -1039,10 +1172,10 @@ def dissolve(
             Can be used to avoid huge geometries being created if the input
             geometries are very interconnected.
             Defaults to 1 (= the output is not tiled).
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
             will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
             the precision. Defaults to 0.0.
@@ -1057,6 +1190,11 @@ def dissolve(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`dissolve_within_distance`: dissolve all feature within the distance
+          specified of each other
+
 
     .. |spatialite_reference_link| raw:: html
 
@@ -1099,15 +1237,14 @@ def dissolve(
 def export_by_bounds(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    bounds: Tuple[float, float, float, float],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    bounds: tuple[float, float, float, float],
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
-    """
-    Export the rows that intersect with the bounds specified.
+    """Export the rows that intersect with the bounds specified.
 
     If ``explodecollections`` is False and the input and output file type is GeoPackage,
     the fid will be preserved. In other cases this will typically not be the case.
@@ -1116,10 +1253,10 @@ def export_by_bounds(
         input_path (PathLike): the input file
         output_path (PathLike): the file to write the result to
         bounds (Tuple[float, float, float, float]): the bounds to filter on.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -1128,6 +1265,13 @@ def export_by_bounds(
             Defaults to False.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`export_by_distance`: export features that are within a certain distance
+          of features of another layer
+        * :func:`export_by_location`: export features that e.g. intersect with features
+          of another layer
+
     """
     logger = logging.getLogger("geofileops.export_by_bounds")
     logger.info(f"Start, on {input_path}")
@@ -1147,17 +1291,16 @@ def isvalid(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]", None] = None,
     only_invalid: bool = True,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     validate_attribute_data: bool = False,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ) -> bool:
-    """
-    Checks for all geometries in the geofile if they are valid.
+    """Checks for all geometries in the geofile if they are valid.
 
     The results are written to the output file.
 
@@ -1171,10 +1314,10 @@ def isvalid(
             input file. Defaults to None.
         only_invalid (bool, optional): if True, only put invalid results in the
             output file. Deprecated: always treated as True.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -1194,6 +1337,10 @@ def isvalid(
 
     Returns:
         bool: True if all geometries were valid.
+
+    See Also:
+        * :func:`make_valid`: make the geometries in the input layer valid
+
     """
     # Check parameters
     if output_path is not None:
@@ -1224,22 +1371,21 @@ def isvalid(
 def makevalid(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Union[str, None, GeometryType] = None,
+    force_output_geometrytype: str | None | GeometryType = None,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
-    precision: Optional[float] = None,
+    where_post: str | None = None,
+    precision: float | None = None,
     validate_attribute_data: bool = False,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Makes all geometries in the input file valid.
+    """Makes all geometries in the input file valid.
 
     Writes the result to the output path.
 
@@ -1253,10 +1399,10 @@ def makevalid(
     Args:
         input_path (PathLike): The input file.
         output_path (PathLike): The file to write the result to.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -1286,6 +1432,9 @@ def makevalid(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`isvalid`: check if the geometries in the input layer are valid
 
     .. |spatialite_reference_link| raw:: html
 
@@ -1358,17 +1507,16 @@ def makevalid(
 def warp(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    gcps: List[Tuple[float, float, float, float, Optional[float]]],
+    gcps: list[tuple[float, float, float, float, float | None]],
     algorithm: str = "polynomial",
-    order: Optional[int] = None,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    order: int | None = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
-    """
-    Warp all input features to the output file according to the gcps specified.
+    """Warp all input features to the output file according to the gcps specified.
 
     Alternative names:
         - rubbersheet, rubbersheeting
@@ -1385,10 +1533,10 @@ def warp(
             Defaults to "polynomial".
         order (int, optional): if algorithm is "polynomial", the order of the
             polynomial to use for warping.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -1418,20 +1566,19 @@ def select(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     sql_stmt: str,
-    sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = "SQLITE",
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    sql_dialect: Literal["SQLITE", "OGRSQL"] | None = "SQLITE",
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Union[GeometryType, str, None] = None,
+    force_output_geometrytype: GeometryType | str | None = None,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = True,
     nb_parallel: int = 1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    '''
-    Execute a SELECT SQL statement on the input file.
+    '''Execute a SELECT SQL statement on the input file.
 
     The ``sql_stmt`` must be in SQLite dialect and can contain placeholders that will be
     replaced automatically. More details can be found in the notes and examples below.
@@ -1444,10 +1591,10 @@ def select(
         sql_stmt (str): the SELECT SQL statement to execute
         sql_dialect (str, optional): the SQL dialect to use. If None, the default SQL
             dialect of the underlying source is used. Defaults to "SQLITE".
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain, if
             {columns_to_select_str} is used. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
@@ -1475,6 +1622,10 @@ def select(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s). Defaults to False.
 
+    See Also:
+        * :func:`select_two_layers`: select features using two input layers based on a
+          SQL query
+
     Notes:
         By convention, the sqlite query can contain following placeholders that
         will be automatically replaced for you:
@@ -1494,8 +1645,6 @@ def select(
 
         .. code-block:: python
 
-            import geofileops as gfo
-
             minimum_area = 100
             sql_stmt = f"""
                 SELECT ST_Buffer({{geometrycolumn}}, 1) AS {{geometrycolumn}}
@@ -1506,8 +1655,8 @@ def select(
                    AND ST_Area({{geometrycolumn}}) > {minimum_area}
             """
             gfo.select(
-                input_path=...,
-                output_path=...,
+                input_path="input.gpkg",
+                output_path="output.gpkg",
                 sql_stmt=sql_stmt,
             )
 
@@ -1582,21 +1731,20 @@ def simplify(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     tolerance: float,
-    algorithm: Union[str, SimplifyAlgorithm] = "rdp",
+    algorithm: str | SimplifyAlgorithm = "rdp",
     lookahead: int = 8,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Applies a simplify operation on geometry column of the input file.
+    """Applies a simplify operation on geometry column of the input file.
 
     The result is written to the output file specified.
 
@@ -1620,10 +1768,10 @@ def simplify(
                 * **"vw"**: Visvalingam Whyatt: tolerance is an area.
 
         lookahead (int, optional): used for Lang algorithms. Defaults to 8.
-        input_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
-        output_layer (str, optional): input layer name. Optional if the input
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         columns (List[str], optional): list of columns to retain. If None, all standard
             columns are retained. In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
@@ -1702,19 +1850,19 @@ def clip(
     input_path: Union[str, "os.PathLike[Any]"],
     clip_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input_layer: Optional[str] = None,
-    input_columns: Optional[List[str]] = None,
-    clip_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    input_layer: str | None = None,
+    input_columns: list[str] | None = None,
+    clip_layer: str | None = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
+    subdivide_coords: int = 7500,
     force: bool = False,
 ):
-    """
-    Clip the input layer with the clip layer.
+    """Clip the input layer with the clip layer.
 
     The resulting layer will contain the parts of the geometries in the
     input layer that overlap with the dissolved geometries in the clip layer.
@@ -1741,16 +1889,16 @@ def clip(
         input_path (PathLike): The file to clip.
         clip_path (PathLike): The file with the geometries to clip with.
         output_path (PathLike): the file to write the result to
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input_layer (str, optional): input layer name. If None, ``input_path`` should
+            contain only one layer. Defaults to None.
         input_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
             that the "fid" will be aliased eg. to "fid_1". Defaults to None.
-        clip_layer (str, optional): clip layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): output layer name. Optional if the
-            file only contains one layer.
+        clip_layer (str, optional): clip layer name. If None, ``clip_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         explodecollections (bool, optional): True to convert all multi-geometries to
             singular ones after the dissolve. Defaults to False.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
@@ -1765,8 +1913,16 @@ def clip(
             batch. A smaller batch size, possibly in combination with a
             smaller ``nb_parallel``, will reduce the memory usage.
             Defaults to -1: (try to) determine optimal size automatically.
-        force (bool, optional): overwrite existing output file(s).
+        subdivide_coords (int, optional): the input geometries will be subdivided to
+            parts with about ``subdivide_coords`` coordinates during processing which
+            can offer a large speed up for complex geometries. Subdividing can result in
+            extra collinear points being added to the boundaries of the output. If 0, no
+            subdividing is applied. Defaults to 7500.
+        force (bool, optional): True to overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`clip_by_geometry`: clip the input layer by a geometry specified
 
     .. |spatialite_reference_link| raw:: html
 
@@ -1796,57 +1952,60 @@ def clip(
     )
 
 
-def erase(
-    input_path: Union[str, "os.PathLike[Any]"],
-    erase_path: Union[str, "os.PathLike[Any]", None],
+def difference(
+    input1_path: Union[str, "os.PathLike[Any]"],
+    input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
-    input_layer: Optional[str] = None,
-    input_columns: Optional[List[str]] = None,
-    erase_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
+    input2_layer: str | None = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 2000,
     force: bool = False,
 ):
-    """
-    Erase all features in the erase layer from the features in the input layer.
+    """Calculate the difference of the input1 layer and input2 layer.
 
     Notes:
         - Every row in the input layer will result in maximum one row in the
           output layer.
-        - The output will contain the columns from the input layer, no columns from the
-          erase layer. The attribute values wont't be changed, so columns like area,...
+        - The output will contain the columns from the 1st no columns from the 2nd
+          layer. The attribute values wont't be changed, so columns like area,...
           will have to be recalculated manually.
-        - If ``erase_path`` is None, the 1st input layer is used for both inputs but
+        - If ``input2_path`` is None, the 1st input layer is used for both inputs but
           interactions between the same rows in this layer will be ignored. The output
           will be the (pieces of) features in this layer that don't have any
           intersections with other features in this layer.
+        - To speed up processing, complex input geometries are subdivided by default.
+          For these geometries, the output geometries will contain extra collinear
+          points where the subdividing occured. This behaviour can be controlled via the
+          ``subdivide_coords`` parameter.
 
     Alternative names:
-        - QGIS: difference
+        - ArcMap: erase
 
     Args:
-        input_path (PathLike): The file to erase from.
-        erase_path (PathLike, optional): The file with the geometries to erase with. If
-            None, the 1st input layer is used for both inputs but interactions between
-            the same rows in this layer will be ignored. The output will be the (pieces
-            of) features in this layer that don't have any intersections with other
-            features in this layer.
+        input1_path (PathLike): The file to remove/difference from.
+        input2_path (PathLike, optional): The file with the geometries to remove from
+            input1. If None, the 1st input layer is used for both inputs but interactions
+            between the same rows in this layer will be ignored. The output will be the
+            (pieces of) features in this layer that don't have any intersections with
+            other features in this layer.
         output_path (PathLike): the file to write the result to.
-        input_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        input_columns (List[str], optional): list of columns to retain. If None, all
+        input1_layer (str, optional): input layer name. If None, ``input1_path`` should
+            contain only one layer. Defaults to None.
+        input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
             that the "fid" will be aliased eg. to "fid_1". Defaults to None.
-        erase_layer (str, optional): erase layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): output layer name. Optional if the
-            file only contains one layer.
+        input2_layer (str, optional): input2 layer name. If None, ``input2_path`` should
+            contain only one layer. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         explodecollections (bool, optional): True to convert all multi-geometries to
             singular ones after the dissolve. Defaults to False.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
@@ -1869,31 +2028,78 @@ def erase(
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
 
+    See Also:
+        * :func:`identity`: calculate the identity of two layers
+        * :func:`intersection`: calculate the intersection of two layers
+        * :func:`symmetric_difference`: calculate the symmetric difference of two layers
+        * :func:`union`: calculate the union of two layers
+
     .. |spatialite_reference_link| raw:: html
 
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
     """  # noqa: E501
-    logger = logging.getLogger("geofileops.erase")
-    logger.info(f"Start, on {input_path} with {erase_path} to {output_path}")
+    logger = logging.getLogger("geofileops.difference")
+    logger.info(f"Start, on {input1_path} with {input2_path} to {output_path}")
 
-    # In erase_path is None, we are doing a self-overlay
+    # If input2_path is None, we are doing a self-overlay
     overlay_self = False
-    if erase_path is None:
-        if erase_layer is not None:
-            raise ValueError("erase_layer must be None if erase_path is None")
-        erase_path = input_path
-        erase_layer = input_layer
+    if input2_path is None:
+        if input2_layer is not None:
+            raise ValueError("input2_layer must be None if input2_path is None")
+        input2_path = input1_path
+        input2_layer = input1_layer
         overlay_self = True
 
-    return _geoops_sql.erase(
-        input_path=Path(input_path),
-        erase_path=Path(erase_path),
+    return _geoops_sql.difference(
+        input1_path=Path(input1_path),
+        input2_path=Path(input2_path),
         output_path=Path(output_path),
         overlay_self=overlay_self,
-        input_layer=input_layer,
-        input_columns=input_columns,
-        erase_layer=erase_layer,
+        input1_layer=input1_layer,
+        input1_columns=input1_columns,
+        input2_layer=input2_layer,
+        output_layer=output_layer,
+        explodecollections=explodecollections,
+        gridsize=gridsize,
+        where_post=where_post,
+        nb_parallel=nb_parallel,
+        batchsize=batchsize,
+        subdivide_coords=subdivide_coords,
+        force=force,
+    )
+
+
+def erase(
+    input_path: Union[str, "os.PathLike[Any]"],
+    erase_path: Union[str, "os.PathLike[Any]", None],
+    output_path: Union[str, "os.PathLike[Any]"],
+    input_layer: str | None = None,
+    input_columns: list[str] | None = None,
+    erase_layer: str | None = None,
+    output_layer: str | None = None,
+    explodecollections: bool = False,
+    gridsize: float = 0.0,
+    where_post: str | None = None,
+    nb_parallel: int = -1,
+    batchsize: int = -1,
+    subdivide_coords: int = 2000,
+    force: bool = False,
+):
+    """DEPRECATED: please use difference."""
+    warnings.warn(  # pragma: no cover
+        "erase is deprecated because it was renamed to difference. "
+        "Will be removed in a (distant) future version",
+        FutureWarning,
+        stacklevel=2,
+    )
+    return difference(
+        input1_path=input_path,
+        input2_path=erase_path,
+        output_path=output_path,
+        input1_layer=input_layer,
+        input1_columns=input_columns,
+        input2_layer=erase_layer,
         output_layer=output_layer,
         explodecollections=explodecollections,
         gridsize=gridsize,
@@ -1910,21 +2116,20 @@ def export_by_location(
     input_to_compare_with_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     spatial_relations_query: str = "intersects is True",
-    min_area_intersect: Optional[float] = None,
-    area_inters_column_name: Optional[str] = None,
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
-    input2_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    min_area_intersect: float | None = None,
+    area_inters_column_name: str | None = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
+    input2_layer: str | None = None,
+    output_layer: str | None = None,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 7500,
     force: bool = False,
 ):
-    """
-    Exports all features filtered by the specified spatial query.
+    """Exports all features filtered by the specified spatial query.
 
     All features in ``input_to_select_from_path`` that comply to the
     ``spatial_relations_query`` compared with the features in
@@ -1939,7 +2144,7 @@ def export_by_location(
     Some examples of valid ``spatial_relations_query`` values:
 
         - "touches is True or within is True"
-        - "intersect is True and touches is False"
+        - "intersects is True and touches is False"
         - "(T*T***T** is True or 1*T***T** is True) and T*****FF* is False"
 
 
@@ -1956,17 +2161,19 @@ def export_by_location(
             Defaults to None.
         area_inters_column_name (str, optional): column name of the intersect
             area. If None, no area column is added. Defaults to None.
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input1_layer (str, optional): 1st input layer name. If None,
+            ``input_to_select_from_path`` should only contain one layer.
+            Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
             that the "fid" will be aliased eg. to "fid_1". Defaults to None.
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input2_layer (str, optional): 2nd input layer name. If None,
+            ``input_to_compare_with_path`` should contain only one layer.
+            Defaults to None.
         input2_columns (List[str], optional): NA.
-        output_layer (str, optional): output layer name. Optional if the
-            file only contains one layer.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
             will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
             the precision. Defaults to 0.0.
@@ -1985,6 +2192,16 @@ def export_by_location(
             geometries. If 0, no subdividing is applied. Defaults to 7.500.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`export_by_bounds`: export features that intersect with the bounds
+          specified
+        * :func:`export_by_distance`: export features that are within a certain distance
+          of features of another layer
+        * :func:`export_by_location`: export features that e.g. intersect with features
+          of another layer
+        * :func:`join_by_location`: join features that e.g. intersect with features of
+          another layer
 
     .. |spatialite_reference_link| raw:: html
 
@@ -2025,18 +2242,17 @@ def export_by_distance(
     input_to_compare_with_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     max_distance: float,
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
-    input2_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
+    input2_layer: str | None = None,
+    output_layer: str | None = None,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    Exports all features within the distance specified.
+    """Exports all features within the distance specified.
 
     Features in ``input_to_select_from_path`` that are within the distance specified of
     any features in ``input_to_compare_with_path``.
@@ -2046,16 +2262,18 @@ def export_by_distance(
         input_to_compare_with_path (PathLike): the 2nd input file
         output_path (PathLike): the file to write the result to
         max_distance (float): maximum distance
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
+        input1_layer (str, optional): 1st input layer name. If None,
+            ``input_to_select_from_path`` should contain only one layer.
+            Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
             that the "fid" will be aliased eg. to "fid_1". Defaults to None.
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer.
-        output_layer (str, optional): output layer name. Optional if the
-            file only contains one layer.
+        input2_layer (str, optional): 2nd input layer name. If None,
+            ``input_to_compare_with_path`` should contain only one layer.
+            Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
             will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
             the precision. Defaults to 0.0.
@@ -2070,6 +2288,12 @@ def export_by_distance(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`export_by_bounds`: export features that intersect with the bounds
+          specified
+        * :func:`export_by_location`: export features that e.g. intersect with features
+          of another layer
 
     .. |spatialite_reference_link| raw:: html
 
@@ -2103,26 +2327,25 @@ def identity(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 2000,
     force: bool = False,
 ):
-    r"""
-    Calculates the pairwise identity of the two input layers.
+    r"""Calculates the pairwise identity of the two input layers.
 
     The result is the equivalent of the intersection between the two layers + layer 1
-    erased with layer 2.
+    differenced with layer 2.
 
     Notes:
         - The result will contain the attribute columns from both input layers. The
@@ -2130,6 +2353,10 @@ def identity(
           recalculated manually if this is wanted.
         - If ``input2_path`` is None, the 1st input layer is used for both inputs but
           interactions between the same rows in this layer will be ignored.
+        - To speed up processing, complex input geometries are subdivided by default.
+          For these geometries, the output geometries will contain extra collinear
+          points where the subdividing occured. This behaviour can be controlled via the
+          ``subdivide_coords`` parameter.
 
     Args:
         input1_path (PathLike): the 1st input file.
@@ -2137,8 +2364,8 @@ def identity(
             layer is used for both inputs but interactions between the same rows in this
             layer will be ignored.
         output_path (PathLike): the file to write the result to
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2146,8 +2373,8 @@ def identity(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for ``input1_columns``, it is also
             possible to specify "fid". Defaults to None.
@@ -2176,6 +2403,12 @@ def identity(
             subdividing is applied. Defaults to 2000.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`difference`: calculate the difference between two layers
+        * :func:`intersection`: calculate the intersection of two layers
+        * :func:`symmetric_difference`: calculate the symmetric difference of two layers
+        * :func:`union`: calculate the union of two layers
 
     .. |spatialite_reference_link| raw:: html
 
@@ -2220,27 +2453,25 @@ def split(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 2000,
     force: bool = False,
 ):
-    """
-    DEPRECATED: please use identity.
-    """
+    """DEPRECATED: please use identity."""
     warnings.warn(
-        "split() is deprecated because it was renamed to identity(). "
-        "Will be removed in a future version",
+        "split is deprecated because it was renamed to identity. "
+        "Will be removed in a future version.",
         FutureWarning,
         stacklevel=2,
     )
@@ -2272,24 +2503,22 @@ def intersect(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    """
-    DEPRECATED: please use intersection.
-    """
+    """DEPRECATED: please use intersection."""
     warnings.warn(  # pragma: no cover
-        "intersect() is deprecated because it was renamed intersection(). "
+        "intersect is deprecated because it was renamed intersection. "
         "Will be removed in a future version",
         FutureWarning,
         stacklevel=2,
@@ -2317,22 +2546,22 @@ def intersection(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
+    subdivide_coords: int = 15000,
     force: bool = False,
 ):
-    r"""
-    Calculates the pairwise intersection of the two input layers.
+    r"""Calculates the pairwise intersection of the two input layers.
 
     Notes:
         - The result will contain the attribute columns from both input layers. The
@@ -2341,6 +2570,10 @@ def intersection(
         - If ``input2_path`` is None, the 1st input layer is used for both inputs but
           intersections between the same rows in this layer will be omitted from the
           result.
+        - To speed up processing, complex input geometries are subdivided by default.
+          For these geometries, the output geometries will contain extra collinear
+          points where the subdividing occured. This behaviour can be controlled via the
+          ``subdivide_coords`` parameter.
 
     Alternative names:
         - GeoPandas: overlay(how="intersection")
@@ -2351,8 +2584,8 @@ def intersection(
             for both inputs but intersections between the same rows in this layer will
             be omitted from the result.
         output_path (PathLike): the file to write the result to
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2360,8 +2593,8 @@ def intersection(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for ``input1_columns``, it is also
             possible to specify "fid". Defaults to None.
@@ -2383,8 +2616,19 @@ def intersection(
             batch. A smaller batch size, possibly in combination with a
             smaller ``nb_parallel``, will reduce the memory usage.
             Defaults to -1: (try to) determine optimal size automatically.
+        subdivide_coords (int, optional): the input geometries will be subdivided to
+            parts with about ``subdivide_coords`` coordinates during processing which
+            can offer a large speed up for complex geometries. Subdividing can result in
+            extra collinear points being added to the boundaries of the output. If 0, no
+            subdividing is applied. Defaults to 15000.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`difference`: calculate the difference between two layers
+        * :func:`identity`: calculate the identity of two layers
+        * :func:`symmetric_difference`: calculate the symmetric difference of two layers
+        * :func:`union`: calculate the union of two layers
 
     .. |spatialite_reference_link| raw:: html
 
@@ -2420,6 +2664,7 @@ def intersection(
         where_post=where_post,
         nb_parallel=nb_parallel,
         batchsize=batchsize,
+        subdivide_coords=subdivide_coords,
         force=force,
     )
 
@@ -2430,23 +2675,22 @@ def join_by_location(
     output_path: Union[str, "os.PathLike[Any]"],
     spatial_relations_query: str = "intersects is True",
     discard_nonmatching: bool = True,
-    min_area_intersect: Optional[float] = None,
-    area_inters_column_name: Optional[str] = None,
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    min_area_intersect: float | None = None,
+    area_inters_column_name: str | None = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    r"""
-    Joins all features in input1 with all features in input2.
+    r"""Joins all features in input1 with all features in input2.
 
     The output will contain the geometries of input1. The ``spatial_relations_query``
     and ``min_area_intersect`` parameters will determine which geometries of input1 will
@@ -2487,8 +2731,8 @@ def join_by_location(
             to match. Defaults to None.
         area_inters_column_name (str, optional): column name of the intersect
             area. If None no area column is added. Defaults to None.
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2496,15 +2740,15 @@ def join_by_location(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for input1_columns, it is also
             possible to specify "fid". Defaults to None.
         input2_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l2\_".
-        output_layer (str, optional): output layer name. If None, the output_path stem
-            is used. Defaults to None.
+        output_layer (str, optional): output layer name. If None, the ``output_path``
+            stem is used. Defaults to None.
         gridsize (float, optional): the size of the grid the coordinates of the ouput
             will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
             the precision. Defaults to 0.0.
@@ -2519,6 +2763,12 @@ def join_by_location(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`export_by_location`: export features that e.g. intersect with features
+          of another layer
+        * :func:`join_by_distance`: join features that are within a certain distance of
+          features of another layer
 
     .. |spatialite_reference_link| raw:: html
 
@@ -2560,21 +2810,20 @@ def join_nearest(
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     nb_nearest: int,
-    distance: Optional[float] = None,
-    expand: Optional[bool] = None,
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    distance: float | None = None,
+    expand: bool | None = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    r"""
-    Joins features of ``input1`` with the ``nb_nearest`` closest features of ``input2``.
+    r"""Joins features of ``input1`` with the ``nb_nearest`` ones in ``input2``.
 
     In addition to the columns requested via the ``input*_columns`` parameters, the
     following columns will be in the output file as well:
@@ -2608,8 +2857,8 @@ def join_nearest(
         expand (bool): True to keep searching till ``nb_nearest`` items are found. If
             False, only items found within ``distance`` are returned (False is only
             supported if spatialite version >= 5.1 is used).
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2617,8 +2866,8 @@ def join_nearest(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for ``input1_columns``, it is also
             possible to specify "fid". Defaults to None.
@@ -2634,6 +2883,13 @@ def join_nearest(
             Defaults to -1: (try to) determine optimal size automatically.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`export_by_distance`: export features that are within a certain distance
+          of features of another layer
+        * :func:`join_by_location`: join features that e.g. intersect with features of
+          another layer
+
     """
     logger = logging.getLogger("geofileops.join_nearest")
     logger.info(f"select from {input1_path} joined with {input2_path} to {output_path}")
@@ -2663,23 +2919,22 @@ def select_two_layers(
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     sql_stmt: str,
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Optional[GeometryType] = None,
+    force_output_geometrytype: GeometryType | None = None,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = 1,
     batchsize: int = -1,
     force: bool = False,
 ):
-    r'''
-    Execute a SELECT SQL statement on the input files.
+    r'''Execute a SELECT SQL statement on the input files.
 
     The ``sql_stmt`` must be in SQLite dialect and can contain placeholders that will be
     replaced automatically. More details can be found in the notes and examples below.
@@ -2692,8 +2947,8 @@ def select_two_layers(
         output_path (PathLike): the file to write the result to.
         sql_stmt (str): the SELECT SQL statement to be executed. Must be in SQLite
             dialect.
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain if one of the
             {layer1\_columns_...} placeholders is used in ``sql_stmt``. If None, all
             standard columns are retained. In addition to standard columns, it is also
@@ -2702,8 +2957,8 @@ def select_two_layers(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): list of columns to retain if one of the
             {layer2\_columns_...} placeholders is used in ``sql_stmt``. If None is
             specified, all columns are selected. As explained for ``input1_columns``, it
@@ -2766,8 +3021,6 @@ def select_two_layers(
 
         .. code-block:: python
 
-            import geofileops as gfo
-
             minimum_area = 100
             sql_stmt = f"""
                 SELECT layer1.{{input1_geometrycolumn}}
@@ -2781,9 +3034,9 @@ def select_two_layers(
                    AND ST_Area(layer1.{{input1_geometrycolumn}}) > {minimum_area}
             """
             gfo.select_two_layers(
-                input1_path=...,
-                input2_path=...,
-                output_path=...,
+                input1_path="input1.gpkg",
+                input2_path="input2.gpkg",
+                output_path="output.gpkg",
                 sql_stmt=sql_stmt,
             )
 
@@ -2807,9 +3060,12 @@ def select_two_layers(
           {layer1_columns_prefix_str}), they will start with a "," and if no column
           precedes it the SQL statement will be invalid.
 
+    See Also:
+        * :func:`select`: select features from a layer based on a SQL query
+
     Examples:
         An ideal place to get inspiration to write you own advanced queries
-        is in the following source code file: |geofileops_sql_link|.
+        is in the following source code file: |geoops_sql_link|.
 
         Additionally, there are some examples listed here that highlight
         other features/possibilities.
@@ -2855,14 +3111,13 @@ def select_two_layers(
                  WHERE pos = 1
             """
 
-
     .. |spatialite_reference_link| raw:: html
 
         <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
 
-    .. |geofileops_sql_link| raw:: html
+    .. |geoops_sql_link| raw:: html
 
-        <a href="https://github.com/geofileops/geofileops/blob/main/geofileops/util/geofileops_sql.py" target="_blank">geofileops_sql.py</a>
+        <a href="https://github.com/geofileops/geofileops/blob/main/geofileops/util/_geoops_sql.py" target="_blank">_geoops_sql.py</a>
 
     '''  # noqa: E501
     logger = logging.getLogger("geofileops.select_two_layers")
@@ -2893,23 +3148,22 @@ def symmetric_difference(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 2000,
     force: bool = False,
 ):
-    r"""
-    Calculates the pairwise symmetric difference of the two input layers.
+    r"""Calculates the pairwise symmetric difference of the two input layers.
 
     The result will be a layer containing features from both the input and overlay
     layers but with the overlapping areas between the two layers removed.
@@ -2920,6 +3174,10 @@ def symmetric_difference(
           recalculated manually if this is wanted.
         - If ``input2_path`` is None, the 1st input layer is used for both inputs but
           interactions between the same rows in this layer will be ignored.
+        - To speed up processing, complex input geometries are subdivided by default.
+          For these geometries, the output geometries will contain extra collinear
+          points where the subdividing occured. This behaviour can be controlled via the
+          ``subdivide_coords`` parameter.
 
 
     Alternative names:
@@ -2932,8 +3190,8 @@ def symmetric_difference(
           for both inputs but interactions between the same rows in this layer will be
           ignored.
         output_path (PathLike): the file to write the result to
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -2941,8 +3199,8 @@ def symmetric_difference(
             "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for input1_columns, it is also
             possible to specify "fid". Defaults to None.
@@ -2971,6 +3229,12 @@ def symmetric_difference(
             subdividing is applied. Defaults to 2000.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`difference`: calculate the difference between two layers
+        * :func:`identity`: calculate the identity of two layers
+        * :func:`intersection`: calculate the intersection of two layers
+        * :func:`union`: calculate the union of two layers
 
     .. |spatialite_reference_link| raw:: html
 
@@ -3018,23 +3282,22 @@ def union(
     input1_path: Union[str, "os.PathLike[Any]"],
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
-    input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_layer: str | None = None,
+    input1_columns: list[str] | None = None,
     input1_columns_prefix: str = "l1_",
-    input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_layer: str | None = None,
+    input2_columns: list[str] | None = None,
     input2_columns_prefix: str = "l2_",
-    output_layer: Optional[str] = None,
+    output_layer: str | None = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    where_post: Optional[str] = None,
+    where_post: str | None = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
     subdivide_coords: int = 2000,
     force: bool = False,
 ):
-    r"""
-    Calculates the pairwise union of the two input layers.
+    r"""Calculates the pairwise union of the two input layers.
 
     Union needs to be interpreted here as such: the output layer will contain the
     combination of all of the following operations:
@@ -3050,6 +3313,10 @@ def union(
           recalculated manually if this is wanted.
         - If ``input2_path`` is None, the 1st input layer is used for both inputs but
           interactions between the same rows in this layer will be ignored.
+        - To speed up processing, complex input geometries are subdivided by default.
+          For these geometries, the output geometries will contain extra collinear
+          points where the subdividing occured. This behaviour can be controlled via the
+          ``subdivide_coords`` parameter.
 
 
     Alternative names:
@@ -3061,8 +3328,8 @@ def union(
             layer is used for both inputs but interactions between the same rows in this
             layer will be ignored.
         output_path (PathLike): the file to write the result to
-        input1_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input1_layer (str, optional): 1st input layer name. If None, ``input1_path``
+            should contain only one layer. Defaults to None.
         input1_columns (List[str], optional): list of columns to retain. If None, all
             standard columns are retained. In addition to standard columns, it is also
             possible to specify "fid", a unique index available in all input files. Note
@@ -3070,8 +3337,8 @@ def union(
             to "fid_1". Defaults to None.
         input1_columns_prefix (str, optional): prefix to use in the column aliases.
             Defaults to "l1\_".
-        input2_layer (str, optional): input layer name. Optional if the
-            file only contains one layer. Defaults to None.
+        input2_layer (str, optional): 2nd input layer name. If None, ``input2_path``
+            should contain only one layer. Defaults to None.
         input2_columns (List[str], optional): columns to select. If None is specified,
             all columns are selected. As explained for ``input1_columns``, it is also
             possible to specify "fid". Defaults to None.
@@ -3100,6 +3367,12 @@ def union(
             subdividing is applied. Defaults to 2000.
         force (bool, optional): overwrite existing output file(s).
             Defaults to False.
+
+    See Also:
+        * :func:`difference`: calculate the difference between two layers
+        * :func:`identity`: calculate the identity of two layers
+        * :func:`intersection`: calculate the intersection of two layers
+        * :func:`symmetric_difference`: calculate the symmetric difference of two layers
 
     .. |spatialite_reference_link| raw:: html
 
