@@ -1,18 +1,16 @@
-"""
-Module with information about geofile types.
-"""
+"""Module with information about geofile types."""
 
 import ast
 import csv
-from dataclasses import dataclass
 import enum
+from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Dict, List, Union, TYPE_CHECKING
+from typing import TYPE_CHECKING, Any, Union
 
 from osgeo import gdal
 from osgeo_utils.auxiliary.util import GetOutputDriversFor
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     import os
 
 # Enable exceptions for GDAL
@@ -22,25 +20,23 @@ gdal.ogr.UseExceptions()
 
 @dataclass
 class GeofileTypeInfo:
-    """
-    Class with properties of a GeofileType.
-    """
+    """Class with properties of a GeofileType."""
 
     geofiletype: str
     ogrdriver: str
-    suffixes: List[str]
+    suffixes: list[str]
     is_fid_zerobased: bool
     is_spatialite_based: bool
     default_spatial_index: bool
-    suffixes_extrafiles: List[str]
+    suffixes_extrafiles: list[str]
 
 
-geofiletypes: Dict[str, GeofileTypeInfo] = {}
+geofiletypes: dict[str, GeofileTypeInfo] = {}
 
 
 def _init_geofiletypes():
     geofiletypes_path = Path(__file__).resolve().parent / "geofiletypes.csv"
-    with open(geofiletypes_path) as file:
+    with geofiletypes_path.open() as file:
         # Set skipinitialspace to True so the csv can be formatted for readability
         csv.register_dialect("geofiletype_dialect", skipinitialspace=True, strict=True)
         reader = csv.DictReader(file, dialect="geofiletype_dialect")
@@ -74,9 +70,7 @@ def _init_geofiletypes():
 
 
 class GeofileType(enum.Enum):
-    """
-    DEPRECATED Enumeration of relevant geo file types and their properties.
-    """
+    """DEPRECATED Enumeration of relevant geo file types and their properties."""
 
     ESRIShapefile = enum.auto()
     GeoJSON = enum.auto()
@@ -86,8 +80,7 @@ class GeofileType(enum.Enum):
 
     @classmethod
     def _missing_(cls, value):
-        """
-        Expand options in the GeofileType() constructor.
+        """Expand options in the GeofileType() constructor.
 
         Args:
             value (Union[str, int, Driver, Path]):
@@ -102,14 +95,14 @@ class GeofileType(enum.Enum):
 
         def get_geofiletype_for_suffix(suffix: str):
             suffix_lower = suffix.lower()
-            for geofiletype in geofiletypes:
-                if suffix_lower in geofiletypes[geofiletype].suffixes:
+            for geofiletype, geofiletype_info in geofiletypes.items():
+                if suffix_lower in geofiletype_info.suffixes:
                     return GeofileType[geofiletype]
             raise ValueError(f"Unknown extension {suffix}")
 
         def get_geofiletype_for_ogrdriver(ogrdriver: str):
-            for geofiletype in geofiletypes:
-                driver = geofiletypes[geofiletype].ogrdriver
+            for geofiletype, geofiletype_info in geofiletypes.items():
+                driver = geofiletype_info.ogrdriver
                 if driver is not None and driver == ogrdriver:
                     return GeofileType[geofiletype]
             raise ValueError(f"Unknown ogr driver {ogrdriver}")
@@ -149,7 +142,7 @@ class GeofileType(enum.Enum):
         return geofiletypes[self.name].ogrdriver
 
     @property
-    def suffixes_extrafiles(self) -> List[str]:
+    def suffixes_extrafiles(self) -> list[str]:
         """Returns a list of suffixes for the extra files for this GeofileType."""
         return geofiletypes[self.name].suffixes_extrafiles
 
@@ -167,16 +160,14 @@ _init_geofiletypes()
 
 
 class GeofileInfo:
-    """
-    A data object containing meta-information about a geofile.
+    """A data object containing meta-information about a geofile.
 
     Attributes:
         driver (str): the relevant gdal driver for the file.
     """
 
-    def __init__(self, path: Path):
-        """
-        Constructor of Layerinfo.
+    def __init__(self, path: Union[str, "os.PathLike[Any]"]):
+        """Constructor of Layerinfo.
 
         Args:
             path (Path): the path to the file.
@@ -226,7 +217,7 @@ class GeofileInfo:
             return False
 
     @property
-    def suffixes_extrafiles(self) -> List[str]:
+    def suffixes_extrafiles(self) -> list[str]:
         """Returns a list of suffixes for the extra files for this GeofileType."""
         if self.geofiletype_info is not None:
             return self.geofiletype_info.suffixes_extrafiles
@@ -235,23 +226,31 @@ class GeofileInfo:
 
 
 def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
-    """
-    Get the gdal driver name for the file specified.
+    """Get the gdal driver name for the file specified.
 
     Args:
-        path (PathLike): The file path.
+        path (PathLike): The file path. |GDAL_vsi| paths are also supported.
 
     Returns:
         str: The OGR driver name.
-    """
-    path = Path(path)
 
-    def get_driver_for_path(input_path) -> str:
+    .. |GDAL_vsi| raw:: html
+
+        <a href="https://gdal.org/en/stable/user/virtual_file_systems.html" target="_blank">GDAL vsi</a>
+
+    """  # noqa: E501
+    # gdal.OpenEx is relatively slow on windows, so for straightforward cases, avoid it.
+    suffix = Path(path).suffix.lower()
+    if suffix == ".gpkg":
+        return "GPKG"
+    elif suffix == ".shp":
+        return "ESRI Shapefile"
+
+    def get_driver_for_path(input_path: Union[str, "os.PathLike[Any]"]) -> str:
         # If there is no suffix, possibly it is only a suffix, so prefix with filename
-        if input_path.suffix == "":
+        local_path = input_path
+        if Path(input_path).suffix == "":
             local_path = f"temp{input_path}"
-        else:
-            local_path = input_path
 
         drivers = GetOutputDriversFor(local_path, is_raster=False)
         if len(drivers) == 1:
@@ -263,8 +262,7 @@ def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
                 f"Path: {input_path}"
             )
 
-    # If the file exists, determine the driver based on the file.
-    datasource = None
+    # Try to determine the driver by opening the file.
     try:
         datasource = gdal.OpenEx(
             str(path), nOpenFlags=gdal.OF_VECTOR | gdal.OF_READONLY | gdal.OF_SHARED
@@ -272,9 +270,21 @@ def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
         driver = datasource.GetDriver()
         drivername = driver.ShortName
     except Exception as ex:
-        try:
-            drivername = get_driver_for_path(path)
-        except Exception:
+        ex_str = str(ex).lower()
+        if (
+            "no such file or directory" in ex_str
+            or "not recognized as being in a supported file format" in ex_str
+            or "not recognized as a supported file format" in ex_str
+        ):
+            # If the file does not exist or if, for some cases like a csv file,
+            # it is e.g. an empty file that was not recognized yet, try to get the
+            # driver based on only the path.
+            try:
+                drivername = get_driver_for_path(path)
+            except Exception:
+                ex.args = (f"get_driver error for {path}: {ex}",)
+                raise
+        else:
             ex.args = (f"get_driver error for {path}: {ex}",)
             raise
     finally:
@@ -284,8 +294,7 @@ def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
 
 
 def get_geofileinfo(path: Union[str, "os.PathLike[Any]"]) -> GeofileInfo:
-    """
-    Get information about a geofile.
+    """Get information about a geofile.
 
     Not a public function, as the properties are for internal use and might change in
     the future.
@@ -296,4 +305,4 @@ def get_geofileinfo(path: Union[str, "os.PathLike[Any]"]) -> GeofileInfo:
     Returns:
         GeofileInfo: _description_
     """
-    return GeofileInfo(path=Path(path))
+    return GeofileInfo(path=path)
