@@ -1,14 +1,15 @@
-from datetime import datetime
 import logging
+from collections.abc import Iterable
+from datetime import datetime
 from pathlib import Path
-from typing import List, Literal, Optional, Tuple, Union
+from typing import Literal
 
 from pygeoops import GeometryType
 from shapely import wkt
 
 import geofileops as gfo
-from geofileops.util import _ogr_util
-from geofileops.util import _io_util
+from geofileops import LayerInfo
+from geofileops.util import _io_util, _ogr_util
 
 logger = logging.getLogger(__name__)
 
@@ -16,10 +17,10 @@ logger = logging.getLogger(__name__)
 def clip_by_geometry(
     input_path: Path,
     output_path: Path,
-    clip_geometry: Union[Tuple[float, float, float, float], str],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    clip_geometry: tuple[float, float, float, float] | str,
+    input_layer: str | LayerInfo | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -27,6 +28,16 @@ def clip_by_geometry(
     if isinstance(clip_geometry, str):
         geom = wkt.loads(clip_geometry)
         spatial_filter = tuple(geom.bounds)
+
+    force_output_geometrytype = None
+    if not explodecollections:
+        if not isinstance(input_layer, LayerInfo):
+            input_layer = gfo.get_layerinfo(input_path, input_layer)
+        if input_layer.geometrytype is not GeometryType.POINT:
+            # If explodecollections is False and the input type is not point, force the
+            # output type to multi, because clip can cause eg. polygons to be split to
+            # multipolygons.
+            force_output_geometrytype = "PROMOTE_TO_MULTI"
 
     _run_ogr(
         operation="clip_by_geometry",
@@ -38,6 +49,7 @@ def clip_by_geometry(
         output_layer=output_layer,
         columns=columns,
         explodecollections=explodecollections,
+        force_output_geometrytype=force_output_geometrytype,
         force=force,
     )
 
@@ -45,10 +57,10 @@ def clip_by_geometry(
 def export_by_bounds(
     input_path: Path,
     output_path: Path,
-    bounds: Tuple[float, float, float, float],
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    bounds: tuple[float, float, float, float],
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -68,12 +80,12 @@ def export_by_bounds(
 def warp(
     input_path: Path,
     output_path: Path,
-    gcps: List[Tuple[float, float, float, float, Optional[float]]],
+    gcps: list[tuple[float, float, float, float, float | None]],
     algorithm: str = "polynomial",
-    order: Optional[int] = None,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    order: int | None = None,
+    input_layer: str | None = None,
+    output_layer: str | None = None,
+    columns: list[str] | None = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -100,42 +112,42 @@ def _run_ogr(
     operation: str,
     input_path: Path,
     output_path: Path,
-    input_layer: Optional[str] = None,
-    output_layer: Optional[str] = None,
-    input_srs: Union[int, str, None] = None,
-    output_srs: Union[int, str, None] = None,
+    input_layer: str | LayerInfo | None = None,
+    output_layer: str | None = None,
+    input_srs: int | str | None = None,
+    output_srs: int | str | None = None,
     reproject: bool = False,
-    spatial_filter: Optional[Tuple[float, float, float, float]] = None,
-    clip_geometry: Optional[Union[Tuple[float, float, float, float], str]] = None,
-    sql_stmt: Optional[str] = None,
-    sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
+    spatial_filter: tuple[float, float, float, float] | None = None,
+    clip_geometry: tuple[float, float, float, float] | str | None = None,
+    sql_stmt: str | None = None,
+    sql_dialect: Literal["SQLITE", "OGRSQL"] | None = None,
     transaction_size: int = 65536,
     append: bool = False,
     update: bool = False,
     explodecollections: bool = False,
-    force_output_geometrytype: Optional[GeometryType] = None,
+    force_output_geometrytype: GeometryType | str | Iterable[str] | None = None,
     options: dict = {},
-    columns: Optional[List[str]] = None,
-    warp: Optional[dict] = None,
+    columns: list[str] | None = None,
+    warp: dict | None = None,
     force: bool = False,
 ) -> bool:
     # Init
     logger = logging.getLogger(f"geofileops.{operation}")
-    start_time = datetime.now()
-    if input_layer is None:
-        input_layer = gfo.get_only_layer(input_path)
     if _io_util.output_exists(path=output_path, remove_if_exists=force):
         return True
-    if input_layer is None:
-        input_layer = gfo.get_only_layer(input_path)
+
+    start_time = datetime.now()
+    if not isinstance(input_layer, LayerInfo):
+        input_layer = gfo.get_layerinfo(input_path, input_layer, raise_on_nogeom=False)
     if output_layer is None:
         output_layer = gfo.get_default_layer(output_path)
 
     info = _ogr_util.VectorTranslateInfo(
         input_path=input_path,
         output_path=output_path,
-        input_layers=input_layer,
+        input_layers=input_layer.name,
         output_layer=output_layer,
+        access_mode=None,
         input_srs=input_srs,
         output_srs=output_srs,
         reproject=reproject,
@@ -144,8 +156,6 @@ def _run_ogr(
         sql_stmt=sql_stmt,
         sql_dialect=sql_dialect,
         transaction_size=transaction_size,
-        append=append,
-        update=update,
         explodecollections=explodecollections,
         force_output_geometrytype=force_output_geometrytype,
         options=options,
@@ -155,5 +165,5 @@ def _run_ogr(
 
     # Run + return result
     result = _ogr_util.vector_translate_by_info(info)
-    logger.info(f"Ready, took {datetime.now()-start_time}")
+    logger.info(f"Ready, took {datetime.now() - start_time}")
     return result
