@@ -1,20 +1,21 @@
-"""
-Module containing utilities regarding the usage of ogr/gdal functionalities.
-"""
+"""Module containing utilities regarding the usage of ogr/gdal functionalities."""
 
 import logging
 import os
-from pathlib import Path
 import tempfile
+import warnings
+from collections.abc import Iterable
+from pathlib import Path
 from threading import Lock
-from typing import Dict, Iterable, List, Literal, Optional, Tuple, Union
+from typing import Any, Literal, Union
 
-from osgeo import gdal
-from osgeo import ogr
+from osgeo import gdal, ogr
 from pygeoops import GeometryType
 
 import geofileops as gfo
-from geofileops import fileops
+from geofileops import _compat, fileops
+from geofileops.util import _geopath_util, _io_util
+from geofileops.util._general_util import MissingRuntimeDependencyError
 
 # Make sure only one instance per process is running
 lock = Lock()
@@ -29,8 +30,8 @@ class GDALError(Exception):
     def __init__(
         self,
         message: str,
-        log_details: List[str] = [],
-        error_details: List[str] = [],
+        log_details: list[str] = [],
+        error_details: list[str] = [],
     ):
         self.message = message
         self.log_details = log_details
@@ -56,50 +57,77 @@ class GDALError(Exception):
             return super().__str__()
 
 
-ogrtype_to_geometrytype = {
-    ogr.wkbNone: None,
-    ogr.wkbUnknown: GeometryType.GEOMETRY,
-    ogr.wkbPoint: GeometryType.POINT,
-    ogr.wkbLineString: GeometryType.LINESTRING,
-    ogr.wkbPolygon: GeometryType.POLYGON,
-    ogr.wkbTriangle: GeometryType.TRIANGLE,
-    ogr.wkbMultiPoint: GeometryType.MULTIPOINT,
-    ogr.wkbMultiLineString: GeometryType.MULTILINESTRING,
-    ogr.wkbMultiPolygon: GeometryType.MULTIPOLYGON,
-    ogr.wkbGeometryCollection: GeometryType.GEOMETRYCOLLECTION,
-    ogr.wkbPolyhedralSurface: GeometryType.POLYHEDRALSURFACE,
-    ogr.wkbTIN: GeometryType.TIN,
-    ogr.wkbPoint25D: GeometryType.POINTZ,
-    ogr.wkbLineString25D: GeometryType.LINESTRINGZ,
-    ogr.wkbPolygon25D: GeometryType.POLYGONZ,
-    ogr.wkbTriangleZ: GeometryType.TRIANGLEZ,
-    ogr.wkbMultiPoint25D: GeometryType.MULTIPOINTZ,
-    ogr.wkbMultiLineString25D: GeometryType.MULTILINESTRINGZ,
-    ogr.wkbMultiPolygon25D: GeometryType.MULTIPOLYGONZ,
-    ogr.wkbGeometryCollection25D: GeometryType.GEOMETRYCOLLECTIONZ,
-    ogr.wkbPolyhedralSurfaceZ: GeometryType.POLYHEDRALSURFACEZ,
-    ogr.wkbTINZ: GeometryType.TINZ,
-    ogr.wkbPointM: GeometryType.POINTM,
-    ogr.wkbLineStringM: GeometryType.LINESTRINGM,
-    ogr.wkbPolygonM: GeometryType.POLYGONM,
-    ogr.wkbTriangleM: GeometryType.TRIANGLEM,
-    ogr.wkbMultiPointM: GeometryType.MULTIPOINTM,
-    ogr.wkbMultiLineStringM: GeometryType.MULTILINESTRINGM,
-    ogr.wkbMultiPolygonM: GeometryType.MULTIPOLYGONM,
-    ogr.wkbGeometryCollectionM: GeometryType.GEOMETRYCOLLECTIONM,
-    ogr.wkbPolyhedralSurfaceM: GeometryType.POLYHEDRALSURFACEM,
-    ogr.wkbTINM: GeometryType.TINM,
-    ogr.wkbPointZM: GeometryType.POINTZM,
-    ogr.wkbLineStringZM: GeometryType.LINESTRINGZM,
-    ogr.wkbPolygonZM: GeometryType.POLYGONZM,
-    ogr.wkbTriangleZM: GeometryType.TRIANGLEZM,
-    ogr.wkbMultiPointZM: GeometryType.MULTIPOINTZM,
-    ogr.wkbMultiLineStringZM: GeometryType.MULTILINESTRINGZM,
-    ogr.wkbMultiPolygonZM: GeometryType.MULTIPOLYGONZM,
-    ogr.wkbGeometryCollectionZM: GeometryType.GEOMETRYCOLLECTIONZM,
-    ogr.wkbPolyhedralSurfaceZM: GeometryType.POLYHEDRALSURFACEZM,
-    ogr.wkbTINZM: GeometryType.TINZM,
-}
+def spatialite_version_info() -> dict[str, str]:
+    """Returns the versions of the spatialite module used in gdal.
+
+    Versions returned: spatialite_version, geos_version.
+
+    Raises:
+        RuntimeError: if a runtime dependency is not available.
+
+    Returns:
+        Dict[str, str]: a dict with the version of the runtime dependencies.
+    """
+    datasource = None
+    try:
+        test_path = Path(__file__).resolve().parent / "test.gpkg"
+        datasource = gdal.OpenEx(str(test_path))
+        result = datasource.ExecuteSQL("SELECT spatialite_version(), geos_version()")
+        row = result.GetNextFeature()
+        spatialite_version = row.GetField(0)
+        geos_version = row.GetField(1)
+        datasource.ReleaseResultSet(result)
+
+    except Exception as ex:  # pragma: no cover
+        message = f"error getting spatialite_version: {ex}"
+        raise MissingRuntimeDependencyError(message) from ex
+
+    finally:
+        datasource = None
+
+    if not spatialite_version:  # pragma: no cover
+        warnings.warn(
+            "empty gdal spatialite version: probably a geofileops dependency was "
+            "not installed correctly: check the installation instructions in the "
+            "geofileops docs.",
+            stacklevel=1,
+        )
+    if not geos_version:  # pragma: no cover
+        warnings.warn(
+            "empty gdal spatialite GEOS version: probably a geofileops dependency was "
+            "not installed correctly: check the installation instructions in the "
+            "geofileops docs.",
+            stacklevel=1,
+        )
+
+    versions = {
+        "spatialite_version": spatialite_version,
+        "geos_version": geos_version,
+    }
+    return versions
+
+
+def ogrtype_to_name(ogrtype: int | None) -> str:
+    if ogrtype is None:
+        return "NONE"
+    else:
+        geometrytypename = ogr.GeometryTypeToName(ogrtype).replace(" ", "").upper()
+
+    if geometrytypename == "NONE":
+        return geometrytypename
+
+    if geometrytypename == "UNKNOWN(ANY)":
+        return "GEOMETRY"
+
+    if geometrytypename.startswith("3D"):
+        geometrytypename = geometrytypename[2:]
+        geometrytypename = f"{geometrytypename}Z"
+
+    if geometrytypename.startswith("MEASURED"):
+        geometrytypename = geometrytypename[8:]
+        geometrytypename = f"{geometrytypename}M"
+
+    return geometrytypename
 
 
 def get_drivers() -> dict:
@@ -110,9 +138,8 @@ def get_drivers() -> dict:
     return drivers
 
 
-def read_cpl_log(path: Path) -> Tuple[List[str], List[str]]:
-    """
-    Reads a cpl_log file and returns a list with log lines and errors.
+def read_cpl_log(path: Path) -> tuple[list[str], list[str]]:
+    """Reads a cpl_log file and returns a list with log lines and errors.
 
     Args:
         path (Path): the file path to the cpl_log file.
@@ -124,7 +151,7 @@ def read_cpl_log(path: Path) -> Tuple[List[str], List[str]]:
     if not path.exists() or path.stat().st_size == 0:
         return ([], [])
 
-    with open(path) as logfile:
+    with path.open() as logfile:
         log_lines = logfile.readlines()
 
     # Cleanup + check for errors
@@ -140,36 +167,95 @@ def read_cpl_log(path: Path) -> Tuple[List[str], List[str]]:
     return (lines_cleaned, lines_error)
 
 
+def StartTransaction(datasource: gdal.Dataset) -> bool:
+    """Starts a transaction on an open datasource.
+
+    Args:
+        datasource (gdal.Dataset): the datasource to start the transaction on.
+
+    Raises:
+        ValueError: if datasource is None.
+
+    Returns:
+        bool: True if the transaction was started successfully.
+    """
+    if datasource is None:
+        raise ValueError("datasource is None")
+
+    if datasource.TestCapability(ogr.ODsCTransactions):
+        datasource.StartTransaction()
+
+    return True
+
+
+def CommitTransaction(datasource: gdal.Dataset | None) -> bool:
+    """Commits a transaction on an open datasource.
+
+    Args:
+        datasource (gdal.Dataset): the datasource to commit the transaction on. If None,
+            no commit is executed.
+
+    Returns:
+        bool: True if the transaction was committed successfully.
+    """
+    if datasource is None:
+        return False
+
+    if datasource.TestCapability(ogr.ODsCTransactions):
+        datasource.CommitTransaction()
+
+    return True
+
+
+def RollbackTransaction(datasource: gdal.Dataset | None) -> bool:
+    """Rolls back a transaction on an open datasource.
+
+    Args:
+        datasource (gdal.Dataset): the datasource to roll back the transaction on. If
+            None, no rollback is executed.
+
+    Returns:
+        bool: True if the transaction was rolled back successfully.
+    """
+    if datasource is None:
+        return False
+
+    if datasource.TestCapability(ogr.ODsCTransactions):
+        datasource.RollbackTransaction()
+
+    return True
+
+
 class VectorTranslateInfo:
     def __init__(
         self,
-        input_path: Path,
-        output_path: Path,
-        input_layers: Union[List[str], str, None] = None,
-        output_layer: Optional[str] = None,
-        input_srs: Union[int, str, None] = None,
-        output_srs: Union[int, str, None] = None,
+        input_path: Union[str, "os.PathLike[Any]"],
+        output_path: Union[str, "os.PathLike[Any]"],
+        input_layers: list[str] | str | None = None,
+        output_layer: str | None = None,
+        access_mode: str | None = None,
+        input_srs: int | str | None = None,
+        output_srs: int | str | None = None,
         reproject: bool = False,
-        spatial_filter: Optional[Tuple[float, float, float, float]] = None,
-        clip_geometry: Optional[Union[Tuple[float, float, float, float], str]] = None,
-        sql_stmt: Optional[str] = None,
-        sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
-        where: Optional[str] = None,
+        spatial_filter: tuple[float, float, float, float] | None = None,
+        clip_geometry: tuple[float, float, float, float] | str | None = None,
+        sql_stmt: str | None = None,
+        sql_dialect: Literal["SQLITE", "OGRSQL"] | None = None,
+        where: str | None = None,
         transaction_size: int = 65536,
-        append: bool = False,
-        update: bool = False,
         explodecollections: bool = False,
-        force_output_geometrytype: Union[GeometryType, str, None] = None,
+        force_output_geometrytype: GeometryType | str | Iterable[str] | None = None,
         options: dict = {},
-        columns: Optional[Iterable[str]] = None,
-        warp: Optional[dict] = None,
-        preserve_fid: Optional[bool] = None,
-        dst_dimensions: Optional[str] = None,
+        columns: Iterable[str] | None = None,
+        warp: dict | None = None,
+        preserve_fid: bool | None = None,
+        dst_dimensions: str | None = None,
     ):
         self.input_path = input_path
         self.output_path = output_path
         self.input_layers = input_layers
         self.output_layer = output_layer
+        self.access_mode = access_mode
         self.input_srs = input_srs
         self.output_srs = output_srs
         self.reproject = reproject
@@ -179,8 +265,6 @@ class VectorTranslateInfo:
         self.sql_dialect = sql_dialect
         self.where = where
         self.transaction_size = transaction_size
-        self.append = append
-        self.update = update
         self.explodecollections = explodecollections
         self.force_output_geometrytype = force_output_geometrytype
         self.options = options
@@ -196,6 +280,7 @@ def vector_translate_by_info(info: VectorTranslateInfo):
         output_path=info.output_path,
         input_layers=info.input_layers,
         output_layer=info.output_layer,
+        access_mode=info.access_mode,
         input_srs=info.input_srs,
         output_srs=info.output_srs,
         reproject=info.reproject,
@@ -205,8 +290,6 @@ def vector_translate_by_info(info: VectorTranslateInfo):
         sql_dialect=info.sql_dialect,
         where=info.where,
         transaction_size=info.transaction_size,
-        append=info.append,
-        update=info.update,
         explodecollections=info.explodecollections,
         force_output_geometrytype=info.force_output_geometrytype,
         options=info.options,
@@ -218,34 +301,37 @@ def vector_translate_by_info(info: VectorTranslateInfo):
 
 
 def vector_translate(
-    input_path: Union[Path, str],
-    output_path: Path,
-    input_layers: Union[List[str], str, None] = None,
-    output_layer: Optional[str] = None,
-    input_srs: Union[int, str, None] = None,
-    output_srs: Union[int, str, None] = None,
+    input_path: Union[str, "os.PathLike[Any]"],
+    output_path: Union[str, "os.PathLike[Any]"],
+    input_layers: list[str] | str | None = None,
+    output_layer: str | None = None,
+    access_mode: str | None = None,
+    input_srs: int | str | None = None,
+    output_srs: int | str | None = None,
     reproject: bool = False,
-    spatial_filter: Optional[Tuple[float, float, float, float]] = None,
-    clip_geometry: Optional[Union[Tuple[float, float, float, float], str]] = None,
-    sql_stmt: Optional[str] = None,
-    sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
-    where: Optional[str] = None,
+    spatial_filter: tuple[float, float, float, float] | None = None,
+    clip_geometry: tuple[float, float, float, float] | str | None = None,
+    sql_stmt: str | None = None,
+    sql_dialect: Literal["SQLITE", "OGRSQL"] | None = None,
+    where: str | None = None,
     transaction_size: int = 65536,
-    append: bool = False,
-    update: bool = False,
     explodecollections: bool = False,
-    force_output_geometrytype: Union[GeometryType, str, None] = None,
+    force_output_geometrytype: GeometryType | str | Iterable[str] | None = None,
     options: dict = {},
-    columns: Optional[Iterable[str]] = None,
-    warp: Optional[dict] = None,
-    preserve_fid: Optional[bool] = None,
-    dst_dimensions: Optional[str] = None,
+    columns: Iterable[str] | None = None,
+    warp: dict | None = None,
+    preserve_fid: bool | None = None,
+    dst_dimensions: str | None = None,
 ) -> bool:
     # API Doc of VectorTranslateOptions:
-    #   https://gdal.org/api/python/osgeo.gdal.html#osgeo.gdal.VectorTranslateOptions
+    #   https://gdal.org/en/stable/api/python/utilities.html#osgeo.gdal.VectorTranslateOptions
     args = []
-    if isinstance(input_path, str):
-        input_path = Path(input_path)
+    if isinstance(input_path, Path):
+        input_path = input_path.as_posix()
+    if isinstance(columns, str):
+        # If a string is passed, convert to list
+        columns = [columns]
+
     gdal_options = _prepare_gdal_options(options, split_by_option_type=True)
 
     # Input file parameters
@@ -266,10 +352,6 @@ def vector_translate(
         input_srs = f"EPSG:{input_srs}"
 
     # Sql'ing, Filtering, clipping
-    if spatial_filter is not None:
-        args.extend(["-spat"])
-        bounds = [str(coord) for coord in spatial_filter]
-        args.extend(bounds)
     if sql_stmt is not None:
         # If sql_stmt starts with "\n" or "\t" for gpkg or with " " for a shp,
         # VectorTranslate outputs no or an invalid file if the statement doesn't return
@@ -304,8 +386,9 @@ def vector_translate(
             raise ValueError(f"unsupported warp algorithm: {algorithm}")
 
     # Input dataset open options
+    input_open_options = []
     for option_name, value in gdal_options["INPUT_OPEN"].items():
-        args.extend(["-oo", f"{option_name}={value}"])
+        input_open_options.append(f"{option_name}={value!s}")
 
     # Output file parameters
     # Get driver for the output_path
@@ -313,23 +396,15 @@ def vector_translate(
 
     # Shapefiles only can have one layer, and the layer name == the stem of the file
     if output_info.driver == "ESRI Shapefile":
-        output_layer = output_path.stem
+        output_layer = _geopath_util.stem(output_path)
 
     # SRS
     if output_srs is not None and isinstance(output_srs, int):
         output_srs = f"EPSG:{output_srs}"
 
     # Output basic options
-    if output_path.exists() is True:
-        if append is True:
-            args.append("-append")
-        if update is True:
-            args.append("-update")
-
     datasetCreationOptions = []
-    # Output dataset creation options are only applicable if a new output file
-    # will be created
-    if output_path.exists() is False or update is False:
+    if access_mode is None:
         dataset_creation_options = gdal_options["DATASET_CREATION"]
         if output_info.driver == "SQLite":
             # If SQLite file, use the spatialite type of sqlite by default
@@ -339,35 +414,48 @@ def vector_translate(
             datasetCreationOptions.extend([f"{option_name}={value}"])
 
     # Output layer options
-    if explodecollections is True:
+    if explodecollections:
         args.append("-explodecollections")
     output_geometrytypes = []
     if force_output_geometrytype is not None:
         if isinstance(force_output_geometrytype, GeometryType):
             output_geometrytypes.append(force_output_geometrytype.name)
-        else:
+        elif isinstance(force_output_geometrytype, str):
             output_geometrytypes.append(force_output_geometrytype)
-    else:
-        if not explodecollections:
-            output_geometrytypes.append("PROMOTE_TO_MULTI")
+        elif isinstance(force_output_geometrytype, Iterable):
+            for geotype in force_output_geometrytype:
+                if isinstance(geotype, GeometryType):
+                    output_geometrytypes.append(geotype.name)
+                elif isinstance(geotype, str):
+                    output_geometrytypes.append(geotype)
+                else:
+                    raise ValueError(f"invalid type in {force_output_geometrytype=}")
+        else:
+            raise ValueError(f"invalid type for {force_output_geometrytype=}")
+    elif (
+        not explodecollections
+        and input_info.driver == "ESRI Shapefile"
+        and output_info.driver != "ESRI Shapefile"
+    ):
+        # Shapefiles are always reported as singlepart type but can also contain
+        # multiparts geometries, so promote to multi
+        output_geometrytypes.append("PROMOTE_TO_MULTI")
+
     if transaction_size is not None:
         args.extend(["-gt", str(transaction_size)])
     if preserve_fid is None:
         if explodecollections:
             # If explodecollections is specified, explicitly disable fid to avoid errors
             args.append("-unsetFid")
+    elif preserve_fid:
+        args.append("-preserve_fid")
     else:
-        if preserve_fid:
-            args.append("-preserve_fid")
-        else:
-            args.append("-unsetFid")
+        args.append("-unsetFid")
 
-    # Output layer creation options are only applicable if a new layer will be
-    # created
+    # Prepare output layer creation options
     layerCreationOptions = []
-    if output_path.exists() is False or (update is True and append is False):
-        for option_name, value in gdal_options["LAYER_CREATION"].items():
-            layerCreationOptions.extend([f"{option_name}={value}"])
+    for option_name, value in gdal_options["LAYER_CREATION"].items():
+        layerCreationOptions.extend([f"{option_name}={value}"])
 
     # General configuration options
     # Remark: passing them as parameter using --config doesn't work, but they are set as
@@ -390,7 +478,7 @@ def vector_translate(
     # with enable_debug=True nothing is logged. In addition, after
     # gdal.ConfigurePythonLogging is called, the CPL_LOG config setting is ignored.
     if "CPL_LOG" not in config_options:
-        gdal_cpl_log_dir = Path(tempfile.gettempdir()) / "geofileops/gdal_cpl_log"
+        gdal_cpl_log_dir = _io_util.get_tempdir() / "geofileops/gdal_cpl_log"
         gdal_cpl_log_dir.mkdir(parents=True, exist_ok=True)
         fd, gdal_cpl_log = tempfile.mkstemp(suffix=".log", dir=gdal_cpl_log_dir)
         os.close(fd)
@@ -405,14 +493,30 @@ def vector_translate(
 
     # Now we can really get to work
     output_ds = None
+    input_has_geom_attribute = False
+    input_has_geometry_attribute = False
     try:
+        # Till gdal 3.10 datetime columns can be interpreted wrongly with arrow.
+        # Additionally, enabling arrow seems to lead to (rare) random crashes, so
+        # for now, disable it by default.
+        use_arrow_key = "OGR2OGR_USE_ARROW_API"
+        if use_arrow_key not in config_options and use_arrow_key not in os.environ:
+            config_options[use_arrow_key] = False
+
         # Go!
         with set_config_options(config_options):
             # Open input datasource already
-            input_ds = gdal.OpenEx(
-                str(input_path),
-                nOpenFlags=gdal.OF_VECTOR | gdal.OF_READONLY | gdal.OF_SHARED,
-            )
+            try:
+                input_ds = gdal.OpenEx(
+                    str(input_path),
+                    nOpenFlags=gdal.OF_VECTOR | gdal.OF_READONLY | gdal.OF_SHARED,
+                    open_options=input_open_options,
+                )
+            except Exception as ex:
+                if "no such file or directory" in str(ex).lower():
+                    raise FileNotFoundError(f"File not found: {input_path}") from ex
+
+                raise
 
             # If output_srs is not specified and the result has 0 rows, gdal creates the
             # output file without srs.
@@ -449,8 +553,6 @@ def vector_translate(
             # creates an attribute column "geometry" in the output file. To be able to
             # detect this case later on, check here if the input file already has an
             # attribute column "geometry".
-            input_has_geom_attribute = False
-            input_has_geometry_attribute = False
             input_layer = input_ds.GetLayer()
             layer_defn = input_layer.GetLayerDefn()
             for field_idx in range(layer_defn.GetFieldCount()):
@@ -467,7 +569,7 @@ def vector_translate(
             options = gdal.VectorTranslateOptions(
                 options=args_copy,
                 format=output_info.driver,
-                accessMode=None,
+                accessMode=access_mode,
                 srcSRS=input_srs,
                 dstSRS=output_srs,
                 reproject=reproject,
@@ -501,60 +603,8 @@ def vector_translate(
         if output_ds is None:
             raise RuntimeError("output_ds is None")
 
-        # Sometimes an invalid output file is written, so close and try to reopen it.
-        output_ds = None
-        if output_path.exists():
-            try:
-                output_ds = gdal.OpenEx(
-                    str(output_path), nOpenFlags=gdal.OF_VECTOR | gdal.OF_UPDATE
-                )
-
-                # If the (first) output row contains NULL as geom/geometry, gdal will
-                # add an attribute column with the name of the (alias of) the geometry
-                # column: so "geometry" or "geom".
-                # To fix this, delete the "geom" or "geometry" attribute column if
-                # present if the input file didn't have an attribute column with this
-                # name.
-                # Bug documented in https://github.com/geofileops/geofileops/issues/313
-                #
-                # Remark: this check must be done on the reopened output file because
-                # in some cases the "geometrycolumn" is incorrectly listed in the field
-                # list of the Dataset returned by VectorTranslate. E.g. when the input
-                # file is empty.
-                if not input_has_geometry_attribute or not input_has_geom_attribute:
-                    assert isinstance(output_ds, gdal.Dataset)
-                    if output_layer is not None:
-                        result_layer = output_ds.GetLayer(output_layer)
-                    elif output_ds.GetLayerCount() == 1:
-                        result_layer = output_ds.GetLayerByIndex(0)
-                    else:
-                        result_layer = None
-                        logger.warning(
-                            "Unable to determine output layer, so not able to remove "
-                            "possibly incorrect geom and geometry text columns, with "
-                            f"input_path: {input_path}, output_path: {output_path}"
-                        )
-
-                    # Output layer was found, so check it
-                    if result_layer is not None:
-                        layer_defn = result_layer.GetLayerDefn()
-                        for field_idx in range(layer_defn.GetFieldCount()):
-                            name = layer_defn.GetFieldDefn(field_idx).GetName().lower()
-                            if (name == "geom" and not input_has_geom_attribute) or (
-                                name == "geometry" and not input_has_geometry_attribute
-                            ):
-                                result_layer.DeleteField(field_idx)
-                                break
-
-            except Exception as ex:
-                logger.info(
-                    f"Opening output file gave error, probably the input file was "
-                    f"empty, no rows were selected or geom was NULL: {ex}"
-                )
-                gfo.remove(output_path)
-            finally:
-                output_ds = None
-
+    except FileNotFoundError:
+        raise
     except Exception as ex:
         output_ds = None
 
@@ -566,19 +616,27 @@ def vector_translate(
         # Read cpl_log file
         log_lines, log_errors = read_cpl_log(gdal_cpl_log_path)
 
-        # Raise
         raise GDALError(
             message, log_details=log_lines, error_details=log_errors
-        ).with_traceback(ex.__traceback__)
+        ).with_traceback(ex.__traceback__) from None
 
     finally:
         output_ds = None
         input_ds = None
 
+        # Fix/remove invalid files that were written.
+        output_ds = None
+        _validate_file(
+            output_path,
+            output_layer,
+            input_has_geometry_attribute,
+            input_has_geom_attribute,
+        )
+
         if gdal_cpl_log_path.exists():
             # Truncate the cpl log file already, because sometimes it is locked and
             # cannot be unlinked.
-            with open(gdal_cpl_log_path, "r+") as logfile:
+            with gdal_cpl_log_path.open("r+") as logfile:
                 logfile.truncate(0)  # size '0' necessary when using r+
             try:
                 gdal_cpl_log_path.unlink(missing_ok=True)
@@ -588,12 +646,128 @@ def vector_translate(
     return True
 
 
-def _prepare_gdal_options(options: dict, split_by_option_type: bool = False) -> dict:
+def _validate_file(
+    path: Union[str, "os.PathLike[Any]"],
+    layer: str | None,
+    input_has_geometry_attribute: bool,
+    input_has_geom_attribute: bool,
+):
+    """Validate and fix a GPKG file.
+
+    Two things are checked and fixed if needed:
+      - the featurecount of the layer in gpkg_ogr_contents should not be NULL.
+      - when the first row of the input layer has an NULL geometry, a redundant column
+        is sometimes present.
+
+    Args:
+        path (PathLike): the file to check.
+        layer (Optional[str]): the output layer name.
+        input_has_geometry_attribute (bool): True if the input file has a geometry
+            attribute column.
+        input_has_geom_attribute (bool): True if the input file has a geom attribute
+            column.
     """
-    Prepares the options so they are ready to pass on to gdal.
+    if not fileops._vsi_exists(path):
+        return
+
+    try:
+        try:
+            # First try to open file in update mode. If OK, we can fix if needed.
+            output_ds = gdal.OpenEx(
+                str(path), nOpenFlags=gdal.OF_VECTOR | gdal.OF_UPDATE
+            )
+            fix = True
+        except Exception:
+            # Opening in update mode failed, so try to open in read-only mode.
+            output_ds = gdal.OpenEx(
+                str(path), nOpenFlags=gdal.OF_VECTOR | gdal.OF_READONLY
+            )
+            fix = False
+
+        assert isinstance(output_ds, gdal.Dataset)
+
+        # Get the output layer
+        if layer is not None:
+            result_layer = output_ds.GetLayer(layer)
+        elif output_ds.GetLayerCount() == 1:
+            result_layer = output_ds.GetLayerByIndex(0)
+        elif output_ds.GetLayerCount() == 0:
+            # At least a GPKG file with no layers gives issues when opened read-only.
+            # So remove it.
+            # This was raised as an issue in https://github.com/OSGeo/gdal/issues/12284
+            # but apparently there are reasons to keep this behaviour.
+            logger.warning(
+                "Output file has layercount=0, so remove it as opening it read/only"
+                "will lead to an error. Probably the input file was empty, "
+                "no rows were selected, geom was NULL or the SQL was invalid."
+            )
+
+            output_ds = None
+            gfo.remove(path)
+            return False
+
+        else:
+            result_layer = None
+
+        # In some cases output files ended up with NULL featurecount in GDAL < 3.10.1.
+        # This was fixed in https://github.com/OSGeo/gdal/pull/11275, but getting the
+        # featurecount of the layer will fix this.
+        if result_layer is not None and not _compat.GDAL_GTE_3101:
+            result_layer.GetFeatureCount()
+
+        # If the (first) output row contains NULL as geom/geometry, gdal will
+        # add an attribute column with the name of the (alias of) the geometry
+        # column: so "geometry" or "geom".
+        # To fix this, delete the "geom" or "geometry" attribute column if
+        # present if the input file didn't have an attribute column with this
+        # name.
+        # Bug documented in https://github.com/geofileops/geofileops/issues/313
+        #
+        # Remark: this check must be done on the reopened output file because
+        # in some cases the "geometrycolumn" is incorrectly listed in the field
+        # list of the Dataset returned by VectorTranslate. E.g. when the input
+        # file is empty.
+        if not input_has_geometry_attribute or not input_has_geom_attribute:
+            # Output layer was found, so check it
+            if result_layer is not None:
+                layer_defn = result_layer.GetLayerDefn()
+                for field_idx in range(layer_defn.GetFieldCount()):
+                    name = layer_defn.GetFieldDefn(field_idx).GetName().lower()
+                    if (name == "geom" and not input_has_geom_attribute) or (
+                        name == "geometry" and not input_has_geometry_attribute
+                    ):
+                        if fix:
+                            result_layer.DeleteField(field_idx)
+                        else:
+                            return False
+
+                        break
+
+            else:
+                logger.warning(
+                    "Unable to determine output layer, so not able to remove "
+                    "possibly incorrect geom and geometry text columns, with "
+                    f"{path=}"
+                )
+
+    except Exception as ex:
+        # In gdal 3.10, invalid gpkg files are still written when an invalid sql
+        # is used if a new file is created or an existing one is overwritten.
+        logger.warning(
+            "Opening output file gave error, so remove it. Probably the input file was "
+            f"empty, no rows were selected, geom was NULL or the SQL was invalid: {ex}"
+        )
+        gfo.remove(path)
+
+    finally:
+        output_ds = None
+
+
+def _prepare_gdal_options(options: dict, split_by_option_type: bool = False) -> dict:
+    """Prepares the options so they are ready to pass on to gdal.
 
         - Uppercase the option key
-        - Check if the option types are on of the supported ones:
+        - Check if the option types are one of the supported ones:
 
             - LAYER_CREATION: layer creation option (lco)
             - DATASET_CREATION: dataset creation option (dsco)
@@ -621,7 +795,7 @@ def _prepare_gdal_options(options: dict, split_by_option_type: bool = False) -> 
         "DESTINATION_OPEN",
         "CONFIG",
     ]
-    prepared_options: Dict[str, dict] = {
+    prepared_options: dict[str, dict] = {
         option_type: {} for option_type in option_types
     }
 
@@ -652,16 +826,15 @@ def _prepare_gdal_options(options: dict, split_by_option_type: bool = False) -> 
         result = prepared_options
     else:
         result = {}
-        for option_type in prepared_options:
-            for option_name, value in prepared_options[option_type].items():
-                result[f"{option_type}.{option_name}"] = value
+        for option_type_key, option_type_value in prepared_options.items():
+            for option_name, value in option_type_value.items():
+                result[f"{option_type_key}.{option_name}"] = value
 
     return result
 
 
 class set_config_options:
-    """
-    Context manager to set config options.
+    """Context manager to set config options.
 
     Args:
         config_options (dict): dict with config options to set.
@@ -687,6 +860,6 @@ class set_config_options:
     def __exit__(self, type, value, traceback):
         # Remove config options that were set
         # TODO: delete loop + uncomment if SetConfigOptions() is supported
-        for name, value in self.config_options.items():
+        for name, _ in self.config_options.items():
             gdal.SetConfigOption(name, None)
         # gdal.SetConfigOptions(self.config_options_backup)
