@@ -1,42 +1,73 @@
-"""
-Module containing utilities regarding processes.
-"""
-from concurrent import futures
+"""Module containing utilities regarding processes."""
+
+import multiprocessing
+import multiprocessing.context
 import os
-from typing import Optional
+from concurrent import futures
+
 import psutil
+
+WORKER_TYPES = {"threads", "processes"}
 
 
 class PooledExecutorFactory:
-    """
-    Context manager to create an Executor.
+    """Context manager to create a pooled executor.
 
     Args:
-        threadpool (bool, optional): True to get a ThreadPoolExecutor,
-            False to get a ProcessPoolExecutor. Defaults to True.
-        max_workers (int, optional): Max number of workers.
+        worker_type (str, optional): type of executor pool to create.
+            "threads" for a ThreadPoolExecutor, "processes" for a ProcessPoolExecutor.
+        max_workers (int, optional): Maximum number of workers.
             Defaults to None to get automatic determination.
-        initialisze (function, optional): Function that does initialisations.
+        initializer (function, optional): Function that does initialisations.
+        mp_context (BaseContext, optional): multiprocessing context if processes are
+            used. If None, "forkserver" will be used on linux to avoid risks on getting
+            deadlocks. Defaults to None.
+
     """
 
-    def __init__(self, threadpool: bool = True, max_workers=None, initializer=None):
-        self.threadpool = threadpool
+    def __init__(
+        self,
+        worker_type: str = "processes",
+        max_workers=None,
+        initializer=None,
+        mp_context: multiprocessing.context.BaseContext | None = None,
+    ):
+        self.worker_type = worker_type.lower()
+        if self.worker_type not in WORKER_TYPES:
+            raise ValueError(
+                f"Invalid worker_type: {self.worker_type}. "
+                f"Must be one of {WORKER_TYPES}."
+            )
+
         if max_workers is not None and os.name == "nt":
             self.max_workers = min(max_workers, 61)
         else:
             self.max_workers = max_workers
+
         self.initializer = initializer
-        self.pool: Optional[futures.Executor] = None
+        self.mp_context = mp_context
+        if mp_context is None and os.name not in {"nt", "darwin"}:
+            # On linux, overrule default to "forkserver" to avoid risks to deadlocks
+            self.mp_context = multiprocessing.get_context("forkserver")
+        self.pool: futures.Executor | None = None
 
     def __enter__(self) -> futures.Executor:
-        if self.threadpool:
+        if self.worker_type == "threads":
             self.pool = futures.ThreadPoolExecutor(
                 max_workers=self.max_workers, initializer=self.initializer
             )
-        else:
+        elif self.worker_type == "processes":
             self.pool = futures.ProcessPoolExecutor(
-                max_workers=self.max_workers, initializer=self.initializer
+                max_workers=self.max_workers,
+                initializer=self.initializer,
+                mp_context=self.mp_context,
             )
+        else:
+            raise ValueError(
+                f"Invalid worker_type: {self.worker_type}. "
+                f"Must be one of {WORKER_TYPES}."
+            )
+
         return self.pool
 
     def __exit__(self, type, value, traceback):
@@ -44,19 +75,50 @@ class PooledExecutorFactory:
             self.pool.shutdown(wait=True)
 
 
-def initialize_worker():
-    # We don't want the workers to block the entire system, so make them nice
+def initialize_worker(worker_type: str, nice_value: int = 15):
+    """Some default inits.
+
+    Following things are done:
+    - Set the worker process priority low (to `nice_value`) so the workers don't block
+      the system.
+    - Reduce OpenMP threads to avoid committed memory getting very high. You can specify
+      the number of threads using the environment variable `GFO_WORKER_OMP_NUM_THREADS`.
+
+    Args:
+        worker_type (str): The type of worker to initialize.
+            "threads" for thread pool, "processes" for process pool.
+        nice_value (int, optional): The niceness value to set for the worker. 19 is the
+            maximum niceness (lowest priority), and -20 is the minimum
+            (highest priority). Defaults to 15.
+    """
+    worker_type = worker_type.lower()
+    if worker_type not in WORKER_TYPES:
+        raise ValueError(
+            f"Invalid worker_type: {worker_type}. Must be one of {WORKER_TYPES}."
+        )
+
+    if worker_type == "processes":
+        # Reduce OpenMP threads to avoid unnecessary inflated committed memory when
+        # using multiprocessing.
+        # Should work for any numeric library used (openblas, mkl,...).
+        # Ref: https://stackoverflow.com/questions/77764228/pandas-scipy-high-commit-memory-usage-windows
+        worker_omp_threads = os.environ.get("GFO_WORKER_OMP_NUM_THREADS", "1")
+        os.environ["OMP_NUM_THREADS"] = worker_omp_threads
+
+    # We don't want the workers to block the entire system, so make the workers nice
     # if they aren't quite nice already.
-    # Remark: on linux, depending on system settings it is not possible to
-    # decrease niceness, even if it was you who niced before.
-    nice_value = 15
+    #
+    # Remarks:
+    #   - on linux, depending on system settings it is not possible to decrease
+    #     niceness, even if it was you who niced before.
+    #   - we are setting the niceness of the process here, so in case of
+    #     `worker_type=threads` the main thread/process will also be impacted.
     if getprocessnice() < nice_value:
         setprocessnice(nice_value)
 
 
 def getprocessnice() -> int:
-    """
-    Get the niceness of the current process.
+    """Get the niceness of the current process.
 
     The nice value can (typically) range from 19, which gives all other
     processes priority, to -20, which means that this process will take
@@ -77,8 +139,7 @@ def getprocessnice() -> int:
 
 
 def setprocessnice(nice_value: int):
-    """
-    Set the niceness of the current process.
+    """Set the niceness of the current process.
 
     The nice value can (typically) range from 19, which gives all other
     processes priority, to -20, which means that this process will take
