@@ -1,11 +1,10 @@
 # import datetime
 import logging
-from typing import Optional
 
 import pygeoops
 import shapely
-from shapely.geometry.base import BaseGeometry
 import shapely.ops
+from shapely.geometry.base import BaseGeometry
 
 from geofileops.util import _geoseries_util
 
@@ -18,9 +17,8 @@ def gfo_difference_collection(
     geom_to_subtract_wkb: bytes,
     keep_geom_type: int = 0,
     subdivide_coords: int = 2000,
-) -> Optional[bytes]:
-    """
-    Applies the difference of geom_to_subtract on geom.
+) -> bytes | None:
+    """Applies the difference of geom_to_subtract on geom.
 
     If the input geometry has many points, they can be subdivided in smaller parts
     to potentially speed up processing as controlled by parameter `subdivide_coords`.
@@ -53,8 +51,6 @@ def gfo_difference_collection(
         if geom_wkb is None:
             return None
         if geom_to_subtract_wkb is None:
-            return geom_wkb
-        if subdivide_coords <= 0:
             return geom_wkb
 
         # Extract wkb's, and return if empty
@@ -98,9 +94,8 @@ def gfo_difference_collection(
         return None
 
 
-def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> Optional[bytes]:
-    """
-    Reduces the precision of the geometry to the gridsize specified.
+def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> bytes | None:
+    """Reduces the precision of the geometry to the gridsize specified.
 
     If reducing the precison leads to a topologyerror, retries after applying make_valid
     and returns the input if it still fails.
@@ -163,9 +158,8 @@ def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> Optional[bytes]:
 def gfo_split(
     geom_wkb: bytes,
     blade_wkb: bytes,
-) -> Optional[bytes]:
-    """
-    Applies a split in the geom using the blade specified.
+) -> bytes | None:
+    """Applies a split in the geom using the blade specified.
 
     Args:
         geom_wkb (bytes): geometry to substract geom_to_subtract_wkb from in wkb format.
@@ -221,8 +215,7 @@ def gfo_split(
 
 
 def gfo_subdivide(geom_wkb: bytes, coords: int = 2000):
-    """
-    Divide the input geometry to smaller parts using rectilinear lines.
+    """Divide the input geometry to smaller parts using rectilinear lines.
 
     Args:
         geom_wkb (geometry): the geometry to subdivide in wkb format.
