@@ -1,22 +1,23 @@
-"""
-Module to benchmark geofileops operations.
-"""
+"""Module to benchmark geofileops operations."""
 # ruff: noqa: D103
 
-from datetime import datetime
+import inspect
 import logging
 import multiprocessing
-import inspect
+from datetime import datetime
 from pathlib import Path
 
+import shapely
+
+import geofileops as gfo
 from benchmark.benchmarker import RunResult
 from benchmark.benchmarks import testdata
-import geofileops as gfo
-from geofileops.util import _geoops_sql
-from geofileops.util import _geoops_gpd
+from geofileops.util import _geoops_gpd, _geoops_sql
 
 logger = logging.getLogger(__name__)
-nb_parallel = min(multiprocessing.cpu_count(), 12)
+nb_complex_polys = 12
+nb_complex_polys_coords = 100_000
+nb_parallel = min(multiprocessing.cpu_count(), nb_complex_polys)
 
 
 def _get_package() -> str:
@@ -29,7 +30,7 @@ def _get_version() -> str:
 
 def buffer(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -40,7 +41,7 @@ def buffer(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="buffer",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="buffer on agri parcel layer BEFL (~500.000 polygons)",
+        operation_descr=f"buffer on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -51,7 +52,7 @@ def buffer(tmp_dir: Path) -> RunResult:
 
 def buffer_spatialite(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -64,7 +65,7 @@ def buffer_spatialite(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="buffer_spatialite",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="buffer on agri parcel layer BEFL (~500.000 polygons)",
+        operation_descr=f"buffer on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -75,7 +76,7 @@ def buffer_spatialite(tmp_dir: Path) -> RunResult:
 
 def buffer_gridsize_spatialite(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -93,10 +94,8 @@ def buffer_gridsize_spatialite(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="buffer_gridsize_spatialite",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            "buffer with gridsize 0.1 on agri parcel layer BEFL (~500.000 polygons)"
-        ),
-        run_details={"nb_cpu": multiprocessing.cpu_count()},
+        operation_descr=f"buffer with gridsize 0.1 on {input_descr}",
+        run_details={"nb_cpu": nb_parallel},
     )
 
     # Cleanup and return
@@ -106,7 +105,7 @@ def buffer_gridsize_spatialite(tmp_dir: Path) -> RunResult:
 
 def buffer_gpd(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -119,7 +118,7 @@ def buffer_gpd(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="buffer_gpd",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="buffer on agri parcel layer BEFL (~500.000 polygons)",
+        operation_descr=f"buffer on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -130,7 +129,7 @@ def buffer_gpd(tmp_dir: Path) -> RunResult:
 
 def dissolve_nogroupby(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -147,7 +146,7 @@ def dissolve_nogroupby(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="dissolve",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="dissolve on agri parcels BEFL (~500.000 polygons)",
+        operation_descr=f"dissolve on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -158,7 +157,7 @@ def dissolve_nogroupby(tmp_dir: Path) -> RunResult:
 
 def dissolve_groupby(tmp_dir: Path) -> RunResult:
     # Init
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -166,7 +165,7 @@ def dissolve_groupby(tmp_dir: Path) -> RunResult:
     gfo.dissolve(
         input_path,
         output_path,
-        groupby_columns=["GEWASGROEP"],
+        groupby_columns=["GWSGRPH_LB"],
         explodecollections=True,
         nb_parallel=nb_parallel,
         force=True,
@@ -176,9 +175,7 @@ def dissolve_groupby(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation="dissolve_groupby",
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            "dissolve on agri parcels BEFL (~500.000 polygons), groupby=[GEWASGROEP]"
-        ),
+        operation_descr=f"dissolve on {input_descr}, groupby=[GWSGRPH_LB]",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -189,8 +186,10 @@ def dissolve_groupby(tmp_dir: Path) -> RunResult:
 
 def clip(tmp_dir: Path) -> RunResult:
     # Init
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -205,9 +204,50 @@ def clip(tmp_dir: Path) -> RunResult:
     result = RunResult(
         package="geofileops",
         package_version=gfo.__version__,
-        operation="clip",
+        operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="clip between 2 agri parcel layers BEFL (2*~500.000 polygons)",
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr}",
+        run_details={"nb_cpu": nb_parallel},
+    )
+
+    # Cleanup and return
+    # output_path.unlink()
+    return result
+
+
+def clip_agri_complexpoly(tmp_dir: Path) -> RunResult:
+    # Init
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    info1 = gfo.get_layerinfo(input1_path)
+    bbox = shapely.box(*info1.total_bounds).buffer(-10_000, join_style="mitre").bounds
+    crs = info1.crs
+    input2_path, input2_descr = testdata.create_testfile(
+        bbox=bbox,
+        geoms=3,
+        polys_per_geom=4,
+        points_per_poly=30_000,
+        crs=crs,
+        dst_dir=tmp_dir,
+    )
+
+    # Go!
+    start_time = datetime.now()
+    output_path = tmp_dir / f"{input1_path.stem}_clip_{input2_path.stem}.gpkg"
+    gfo.clip(
+        input_path=input1_path,
+        clip_path=input2_path,
+        output_path=output_path,
+        nb_parallel=nb_parallel,
+        force=True,
+    )
+    result = RunResult(
+        package="geofileops",
+        package_version=gfo.__version__,
+        operation=function_name,
+        secs_taken=(datetime.now() - start_time).total_seconds(),
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -220,8 +260,8 @@ def export_by_location_intersects(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
 
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -242,10 +282,7 @@ def export_by_location_intersects(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            "export_by_location_intersects between 2 agri parcel layers BEFL "
-            "(2*~500.000 polygons)"
-        ),
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr} ",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -259,9 +296,17 @@ def export_by_location_intersects_complexpoly(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
 
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, input1_descr = testdata.TestFile.COMPLEX_POLYS.get_file(
-        tmp_dir, nb_points=300_000
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    info1 = gfo.get_layerinfo(input1_path)
+    bbox = shapely.box(*info1.total_bounds).buffer(-10_000, join_style="mitre").bounds
+    crs = info1.crs
+    input2_path, input2_descr = testdata.create_testfile(
+        bbox=bbox,
+        geoms=3,
+        polys_per_geom=4,
+        points_per_poly=300_000,
+        crs=crs,
+        dst_dir=tmp_dir,
     )
 
     # Go!
@@ -283,9 +328,7 @@ def export_by_location_intersects_complexpoly(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            f"{function_name} between agri parcels (~500k poly) and {input1_descr}"
-        ),
+        operation_descr=(f"{function_name} between {input1_descr} and {input2_descr}"),
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -297,39 +340,10 @@ def export_by_location_intersects_complexpoly(tmp_dir: Path) -> RunResult:
 
 def intersection(tmp_dir: Path) -> RunResult:
     # Init
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
-
-    # Go!
-    start_time = datetime.now()
-    output_path = tmp_dir / f"{input1_path.stem}_inters_{input2_path.stem}.gpkg"
-    gfo.intersection(
-        input1_path=input1_path,
-        input2_path=input2_path,
-        output_path=output_path,
-        nb_parallel=nb_parallel,
-        force=True,
-    )
-    result = RunResult(
-        package=_get_package(),
-        package_version=_get_version(),
-        operation="intersection",
-        secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=("intersection between 2 agri parcel layers (2*~500k poly)"),
-        run_details={"nb_cpu": nb_parallel},
-    )
-
-    # Cleanup and return
-    output_path.unlink()
-    return result
-
-
-def intersection_complexpoly_agri(tmp_dir: Path) -> RunResult:
-    # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
 
-    input1_path, input1_descr = testdata.TestFile.COMPLEX_POLYS.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -346,9 +360,85 @@ def intersection_complexpoly_agri(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            f"{function_name} between {input1_descr} and agri parcels (~500k poly)"
-        ),
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr}",
+        run_details={"nb_cpu": nb_parallel},
+    )
+
+    # Cleanup and return
+    output_path.unlink()
+    return result
+
+
+def intersection_complexpoly_agri(tmp_dir: Path) -> RunResult:
+    # Init
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    info1 = gfo.get_layerinfo(input1_path)
+    bbox = shapely.box(*info1.total_bounds).buffer(-10_000, join_style="mitre").bounds
+    crs = info1.crs
+    input2_path, input2_descr = testdata.create_testfile(
+        bbox=bbox,
+        geoms=3,
+        polys_per_geom=4,
+        points_per_poly=30_000,
+        crs=crs,
+        dst_dir=tmp_dir,
+    )
+
+    # Go!
+    start_time = datetime.now()
+    output_path = tmp_dir / f"{input1_path.stem}_inters_{input2_path.stem}.gpkg"
+    gfo.intersection(
+        input1_path=input1_path,
+        input2_path=input2_path,
+        output_path=output_path,
+        nb_parallel=nb_parallel,
+        force=True,
+    )
+    result = RunResult(
+        package=_get_package(),
+        package_version=_get_version(),
+        operation=function_name,
+        secs_taken=(datetime.now() - start_time).total_seconds(),
+        operation_descr=(f"{function_name} between {input1_descr} and {input2_descr}"),
+        run_details={"nb_cpu": nb_parallel},
+    )
+
+    # Cleanup and return
+    output_path.unlink()
+    return result
+
+
+def intersection_complexpoly_complexpoly(tmp_dir: Path) -> RunResult:
+    # Init
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    bbox = (30_000.123, 170_000.123, 250_000, 250_000)
+    input1_path, input1_descr = testdata.create_testfile(
+        bbox=bbox, geoms=3, polys_per_geom=4, points_per_poly=30_000, dst_dir=tmp_dir
+    )
+    bbox = (31_000.123, 171_000.123, 250_000, 250_000)
+    input2_path, input2_descr = testdata.create_testfile(
+        bbox=bbox, geoms=3, polys_per_geom=4, points_per_poly=30_000, dst_dir=tmp_dir
+    )
+
+    # Go!
+    start_time = datetime.now()
+    output_path = tmp_dir / f"{input1_path.stem}_inters_{input2_path.stem}.gpkg"
+    gfo.intersection(
+        input1_path=input1_path,
+        input2_path=input2_path,
+        output_path=output_path,
+        nb_parallel=nb_parallel,
+        force=True,
+    )
+    result = RunResult(
+        package=_get_package(),
+        package_version=_get_version(),
+        operation=function_name,
+        secs_taken=(datetime.now() - start_time).total_seconds(),
+        operation_descr=(f"{function_name} between {input1_descr} and {input2_descr}"),
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -359,12 +449,14 @@ def intersection_complexpoly_agri(tmp_dir: Path) -> RunResult:
 
 def intersection_gridsize(tmp_dir: Path) -> RunResult:
     # Init
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
-    output_path = tmp_dir / f"{input1_path.stem}_inters_grid01_{input2_path.stem}.gpkg"
+    output_path = tmp_dir / f"{input1_path.stem}_inters_grids_{input2_path.stem}.gpkg"
     gfo.intersection(
         input1_path=input1_path,
         input2_path=input2_path,
@@ -376,11 +468,10 @@ def intersection_gridsize(tmp_dir: Path) -> RunResult:
     result = RunResult(
         package=_get_package(),
         package_version=_get_version(),
-        operation="intersection_gridsize",
+        operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
         operation_descr=(
-            "intersection with gridsize 0.001 between 2 agri parcel layers BEFL "
-            "(2*~500k polygons)"
+            f"{function_name} 0.001 between {input1_descr} and {input2_descr}"
         ),
         run_details={"nb_cpu": nb_parallel},
     )
@@ -391,9 +482,11 @@ def intersection_gridsize(tmp_dir: Path) -> RunResult:
 
 
 def join_by_location_intersects(tmp_dir: Path) -> RunResult:
-    # Init-
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
+    # Init
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -412,12 +505,9 @@ def join_by_location_intersects(tmp_dir: Path) -> RunResult:
     result = RunResult(
         package=_get_package(),
         package_version=_get_version(),
-        operation="join_by_location_intersects",
+        operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            "join_by_location_intersects between 2 agri parcel layers BEFL "
-            "(2*~500.000 polygons)"
-        ),
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -430,7 +520,7 @@ def join_by_location_intersects(tmp_dir: Path) -> RunResult:
 def makevalid_gridsize_gpd(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -443,7 +533,7 @@ def makevalid_gridsize_gpd(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=f"{function_name} on agri parcel layer BEFL (~500k polygons)",
+        operation_descr=f"{function_name} on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -455,7 +545,7 @@ def makevalid_gridsize_gpd(tmp_dir: Path) -> RunResult:
 def makevalid_gridsize_spatialite(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -468,7 +558,7 @@ def makevalid_gridsize_spatialite(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=f"{function_name} on agri parcel layer BEFL (~500k polygons)",
+        operation_descr=f"{function_name} on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -480,7 +570,7 @@ def makevalid_gridsize_spatialite(tmp_dir: Path) -> RunResult:
 def makevalid_gpd(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -491,7 +581,7 @@ def makevalid_gpd(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=f"{function_name} on agri parcel layer BEFL (~500k polygons)",
+        operation_descr=f"{function_name} on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -503,7 +593,7 @@ def makevalid_gpd(tmp_dir: Path) -> RunResult:
 def makevalid_spatialite(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
-    input_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input_path, input_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
@@ -514,7 +604,7 @@ def makevalid_spatialite(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=f"{function_name} on agri parcel layer BEFL (~500k polygons)",
+        operation_descr=f"{function_name} on {input_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -527,12 +617,22 @@ def symmetric_difference_complexpolys_agri(tmp_dir: Path) -> RunResult:
     # Init
     function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
 
-    input1_path, input1_descr = testdata.TestFile.COMPLEX_POLYS.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    info2 = gfo.get_layerinfo(input2_path)
+    bbox = shapely.box(*info2.total_bounds).buffer(-10_000, join_style="mitre").bounds
+    crs = info2.crs
+    input1_path, input1_descr = testdata.create_testfile(
+        bbox=bbox,
+        geoms=3,
+        polys_per_geom=4,
+        points_per_poly=30_000,
+        crs=crs,
+        dst_dir=tmp_dir,
+    )
 
     # Go!
     start_time = datetime.now()
-    output_path = tmp_dir / f"{input1_path.stem}_inters_{input2_path.stem}.gpkg"
+    output_path = tmp_dir / f"{input1_path.stem}_symmdiff_{input2_path.stem}.gpkg"
     gfo.symmetric_difference(
         input1_path=input1_path,
         input2_path=input2_path,
@@ -545,9 +645,7 @@ def symmetric_difference_complexpolys_agri(tmp_dir: Path) -> RunResult:
         package_version=_get_version(),
         operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr=(
-            f"{function_name} between {input1_descr} and agriparcels BEFL (~500k poly)"
-        ),
+        operation_descr=f"{function_name} between {input1_descr} and {input2_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
@@ -558,12 +656,14 @@ def symmetric_difference_complexpolys_agri(tmp_dir: Path) -> RunResult:
 
 def union(tmp_dir: Path) -> RunResult:
     # Init
-    input1_path, _ = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
-    input2_path, _ = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
+    function_name = inspect.currentframe().f_code.co_name  # type: ignore[union-attr]
+
+    input1_path, input1_descr = testdata.TestFile.AGRIPRC_2018.get_file(tmp_dir)
+    input2_path, input2_descr = testdata.TestFile.AGRIPRC_2019.get_file(tmp_dir)
 
     # Go!
     start_time = datetime.now()
-    output_path = tmp_dir / f"{input1_path.stem}_inters_{input2_path.stem}.gpkg"
+    output_path = tmp_dir / f"{input1_path.stem}_union_{input2_path.stem}.gpkg"
     gfo.union(
         input1_path=input1_path,
         input2_path=input2_path,
@@ -574,9 +674,9 @@ def union(tmp_dir: Path) -> RunResult:
     result = RunResult(
         package=_get_package(),
         package_version=_get_version(),
-        operation="union",
+        operation=function_name,
         secs_taken=(datetime.now() - start_time).total_seconds(),
-        operation_descr="union between 2 agri parcel layers BEFL (2*~500.000 polygons)",
+        operation_descr=f"{function_name} between {input2_descr} and {input2_descr}",
         run_details={"nb_cpu": nb_parallel},
     )
 
